### 2021-12-03 #3978 @whedon
[PRE REVIEW]: patRoon 2.0: Improved non-target analysis workflows including automated transformation product screening
**Submitting author:** @rickhelmus (<a href="http://orcid.org/0000-0001-9401-3133">Rick Helmus</a>)
**Repository:** <a href="https://github.com/rickhelmus/patRoon" target ="_blank">https://github.com/rickhelmus/patRoon</a>
**Version:** 2.0.0
