### 2021-12-03 #3978 @kthyng
Hi @rickhelmus and thanks for your submission! I see your paper for version 1.0 — can you describe what the review process for that paper entailed? Was there a software review? I know that you are seeking to publish on the basis of improvements between the two versions, but it would still be helpful to know what review went into v1.0. Thanks!
