### 2021-12-04 #3978 @rickhelmus
Dear @kthyng,

Thanks for the swift response!

The previous paper provided a detailed overview of the functionality, implementation, infrastructure, benchmarks and a brief usage overview of patRoon 1.0. While most reviewers primarily focused on the contents of the manuscript, the website, inclusion of unit tests and software installation were also covered (for instance, providing Docker images with patRoon resulted from feedback of a reviewer). However, based on my impressions, I expect that JOSS reviews put more weight on the code quality, documentation, infrastructure and other software related issues.

Hope that helps and I am happy to give more details if needed!
