### 2021-12-04 #3978 @rickhelmus
> @rickhelmus if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, this list of people have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

Regarding any potential reviewers, I came up with the following suggestions:

- jorainer
- lgatto*
- sgibb
- Zhiqiang-PANG
- jsychong
- wesleyburr*
- hechth*
- joewandy*

My selection was mainly based on experience with mass spectrometry and general experience in software development. People noted with a * have previous JOSS contributions.

Thanks,
Rick
