### 2021-12-04 #3978 @schymane
> Hi @rickhelmus and thanks for your submission! I see your paper for version 1.0 — can you describe what the review process for that paper entailed? Was there a software review? I know that you are seeking to publish on the basis of improvements between the two versions, but it would still be helpful to know what review went into v1.0. Thanks!

Thanks @kthyng - in addition to @rickhelmus 's response, I'm adding in the link to the article plus JCheminf requirements for software articles, FYI:

https://jcheminf.biomedcentral.com/articles/10.1186/s13321-020-00477-w

https://jcheminf.biomedcentral.com/submission-guidelines/preparing-your-manuscript/software
