+++ NAMESPACE
@@ -188,6 +188,7 @@ exportMethods(generateComponentsRAMClustR)
 exportMethods(generateComponentsSpecClust)
 exportMethods(generateComponentsTPs)
 exportMethods(generateCompounds)
+exportMethods(generateCompoundsLibrary)
 exportMethods(generateCompoundsMetFrag)
 exportMethods(generateCompoundsSIRIUS)
 exportMethods(generateFormulas)
