+++ DESCRIPTION
@@ -177,6 +177,7 @@ Collate:
     'feature_groups-screening-set.R'
     'components-tps.R'
     'compounds-cluster.R'
+    'compounds-library.R'
     'compounds-metfrag.R'
     'utils-feat_annotations-set.R'
     'compounds-set.R'
