### 2021-12-08 #3978 @danielskatz
Thanks @rickhelmus and @schymane - I'm going to ask the JOSS editors to take a look at this to decide if we will review it - you should hear back in a week or two.
