### 2021-12-11 23fcf7b
handle NA ion_formula from compoundsLib fragInfos, by e.g. annotated column in annotatedPeakList
