+++ R/compounds-set.R
@@ -116,8 +116,7 @@ setMethod("plotSpectrum", "compoundsSet", function(obj, index, groupName, MSPeak
         # UNDONE: this will overwrite consensus algo if present, OK?
         specs <- lapply(specs, function(x)
         {
-            if (!is.null(x[["ion_formula"]]))
-                x[!is.na(ion_formula), mergedBy := set]
+            x[annotated == TRUE, mergedBy := set]
             return(x)
         })
         
