+++ R/utils-compounds.R
@@ -22,16 +22,21 @@ mergeFragInfo <- function(fiLeft, fiRight, leftName, rightName)
     else if (nrow(fiRight) > 0)
     {
         # for overlap: just add label
-        fiLeft <- merge(fiLeft, fiRight[, c("PLID", "mergedBy"), with = FALSE], all.x = TRUE, by = "PLID")
+        fiLeft <- merge(fiLeft, fiRight[, c("PLID", "mergedBy", "ion_formula", "neutral_loss"), with = FALSE], all.x = TRUE, by = "PLID")
         fiLeft[is.na(mergedBy.y), mergedBy := mergedBy.x]
         fiLeft[is.na(mergedBy.x), mergedBy := mergedBy.y]
         fiLeft[!is.na(mergedBy.x) & !is.na(mergedBy.y), mergedBy := paste(mergedBy.x, mergedBy.y, sep = ",")]
         fiLeft[, c("mergedBy.x", "mergedBy.y") := NULL]
+        fiLeft[, ion_formula := fifelse(!is.na(ion_formula.x), ion_formula.x, ion_formula.y)]
+        fiLeft[, neutral_loss := fifelse(!is.na(neutral_loss.x), neutral_loss.x, neutral_loss.y)]
+        fiLeft[, c("ion_formula.x", "ion_formula.y", "neutral_loss.x", "neutral_loss.y") := NULL]
 
         # add unique
         fiUnique <- fiRight[!PLID %in% fiLeft$PLID]
         if (nrow(fiUnique) > 0)
             fiLeft <- rbind(fiLeft, fiUnique, fill = TRUE)
+        
+        setorderv(fiLeft, "PLID")
     }
 
     return(fiLeft)
