+++ R/compounds.R
@@ -418,7 +418,7 @@ setMethod("annotatedPeakList", "compounds", function(obj, index, groupName, MSPe
                     inForms <- annPLForms$annotated
                     
                     # UNDONE: handle different formula assignments?
-                    ret[is.na(ion_formula), ion_formula := annPLForms$ion_formula]
+                    ret[, ion_formula := fifelse(!is.na(ion_formula), ion_formula, annPLForms$ion_formula)]
                     ret[, mergedBy := fcase(inMe & inForms, paste0(algorithm(obj), ",", algorithm(formulas)),
                                             inMe, algorithm(obj),
                                             inForms, algorithm(formulas))]
