### 2021-12-12 067c5ac
assign some missing data
