### 2021-12-12 487cf7d
spectra pre-treatment [skip ci]
