### 2021-12-12 362b9be
various fixes, try to calculate missing SMILES etc, handle NA babel conversion inputs, various fixes
