+++ R/MSLibrary.R
@@ -73,7 +73,7 @@ setMethod("convertToSuspects", "MSLibrary", function(obj)
                  InChI = "InChI",
                  InChIKey = "InChIKey",
                  Formula = "formula",
-                 Precursor_Type = "adduct",
+                 Precursor_type = "adduct",
                  ExactMass = "neutralMass")
     mapCols <- mapCols[names(mapCols) %in% names(ret)]
     setnames(ret, names(mapCols), mapCols)
@@ -102,8 +102,6 @@ setMethod("export", "MSLibrary", function(obj, type, out)
 
 loadMSPLibrary <- function(file, parseComments = TRUE)
 {
-    # UNDONE calculate missing SMILES/InChI(Keys)
-    
     ac <- checkmate::makeAssertCollection()
     checkmate::assertFileExists(file, "r", add = ac)
     aapply(checkmate::assertFlag, . ~ parseComments, fixed = list(add = ac))
@@ -121,9 +119,9 @@ loadMSPLibrary <- function(file, parseComments = TRUE)
     #     return(ret)
     # }))
     
-    # C++ code sets "NA" as string, convert to NA
+    # C++ code sets "NA" as string, convert to NA. Similarly, library may have 'n/a' markers...
     for (j in seq_along(lib$records))
-        set(lib$records, which(lib$records[[j]] == "NA"), j, NA_character_)
+        set(lib$records, which(lib$records[[j]] %chin% c("NA", "n/a")), j, NA_character_)
     
     # Ensure case of column names used by patRoon are consistent
     chCols <- c("Name", "SMILES", "InChI", "InChIKey", "Formula", "Precursor_type", "Ion_mode")
@@ -152,22 +150,21 @@ loadMSPLibrary <- function(file, parseComments = TRUE)
     # ensure mass data is numeric, but ignore conversion warnings
     suppressWarnings(lib$records[, (numCols) := lapply(.SD, as.numeric), .SDcols = numCols])
     
-    printf("Calculating missing formulas\n")
-    lib$records[is.na(Formula) & !is.na(SMILES), Formula := withProg(.N, FALSE, sapply(SMILES, function(smi)
-    {
-        ret <- convertToFormulaBabel(smi, "smi", FALSE)
-        doProgress()
-        return(fifelse(nzchar(ret), ret, NA_character_))
-    }))]
-
-    printf("Calculating missing exact masses\n")
-    lib$records[is.na(ExactMass) & !is.na(Formula), ExactMass := withProg(.N, FALSE, sapply(Formula, function(form)
-    {
-        ret <- tryCatch(getFormulaMass(form), error = function(...) NA_real_)
-        doProgress()
-        return(ret)
-    }))]
-
+    # UNDONE: this is quite slow, skip for now?
+    # printf("Sanitizing SMILES/InChI values... ")
+    # lib$records[!is.na(SMILES), SMILES := babelConvert(SMILES, "smi", "smi", FALSE)]
+    # lib$records[!is.na(InChI), InChI := babelConvert(InChI, "inchi", "inchi", FALSE)]
+    # printf("Done!\n")
+    
+    lib$records <- convertChemDataIfNeeded(lib$records, destFormat = "smi", destCol = "SMILES",
+                                           fromFormats = "inchi", fromCols = "InChI")
+    lib$records <- convertChemDataIfNeeded(lib$records, destFormat = "inchi", destCol = "InChI",
+                                           fromFormats = "smi", fromCols = "SMILES")
+    lib$records <- convertChemDataIfNeeded(lib$records, destFormat = "inchikey", destCol = "InChIKey",
[…]
