### 2021-12-12 a37d9fd
also load computed SMILES/InChIs
