### 2021-12-14 740e8e0
also check for adducts already in the lib
