### 2021-12-14 b52c231
configurable pottential adducts
