### 2021-12-16 7f9ebef
more adduct error handling [skip ci]
