### 2021-12-17 995e7b6
also handle "N/A" for NAs
