### 2021-12-18 9ff78a4
extend checkPackage for GH subdirs and put call. to FALSE
