### 2021-12-18 346d91d
more consistent checkPackage() GitHub repos
