### 2021-12-18 6f44a6b
calculate SPLASHs
