### 2021-12-18 f5bc15c
merge: verify if SPLASHs are there
