### 2021-12-22 94de9e9
make sure that newProject() always adds suspect annotation when it needs to
