### 2021-12-22 #3978 @KristinaRiemer
Thank you for providing the list of possible reviewers who have both software development and domain experience, @rickhelmus!

@lgatto @wesleyburr @hechth would any of you be able and willing to review this JOSS submission of an R package for mass spectrometry analysis? You can view the software at the repo linked in the first comment.
