### 2021-12-22 0e858b1
Typo in KPIC2 default feature param ranges
