### 2021-12-23 0ea298a
small fixes in determining what is already installed [skip ci]
