### 2021-12-23 a267e17
handbook: use PubChemLite for compounds to avoid dependency on PubChem/MetFrag servers
