+++ appveyor.yml
@@ -27,9 +27,9 @@ install:
             Invoke-WebRequest https://github.com/rickhelmus/patRoonDeps/raw/master/ext/MetFrag2.4.5-CL.jar -OutFile "C:\deps\MetFrag2.4.5-CL.jar"
         }
     - ps: >-
-        if (-Not (Test-Path C:\deps\sirius-4.9.9.zip)) {
-            Invoke-WebRequest https://github.com/boecker-lab/sirius/releases/download/v4.9.9/sirius-4.9.9-win64-headless.zip -OutFile "C:\deps\sirius-4.9.9.zip"
-            7z x -aoa -oC:\deps\ C:\deps\sirius-4.9.9.zip
+        if (-Not (Test-Path C:\deps\sirius-4.9.12.zip)) {
+            Invoke-WebRequest https://github.com/boecker-lab/sirius/releases/download/v4.9.12/sirius-4.9.12-win64-headless.zip -OutFile "C:\deps\sirius-4.9.12.zip"
+            7z x -aoa -oC:\deps\ C:\deps\sirius-4.9.12.zip
         }
     - ps: >-
         if (-Not (Test-Path C:\deps\openbabel311.exe)) {
