### 2021-12-23 27ddfa5
setup rapidjson, initial readMoNAJSON
