+++ src/mslibrary.cpp
@@ -13,18 +13,20 @@
 
 #include "utils.h"
 
-struct MSPSpectrum
+namespace {
+
+struct MSLibSpectrum
 {
     std::vector<double> mzs, intensities;
 };
 
-struct MSPRecord
+struct MSLibRecord
 {
     std::unordered_map<std::string, std::string> values;
-    MSPSpectrum spectrum;
+    MSLibSpectrum spectrum;
 };
 
-bool parseComments(const std::string &comments, const std::string &field, std::string &out)
+bool parseMSPComments(const std::string &comments, const std::string &field, std::string &out)
 {
     const std::string toMatch = '"' + field + '=';
     auto start = comments.find(toMatch);
@@ -41,12 +43,46 @@ bool parseComments(const std::string &comments, const std::string &field, std::s
     return false;
 }
 
+Rcpp::List convertRecordsToRData(const std::vector<MSLibRecord> &records, const std::vector<std::string> &keys)
+{
+    Rcpp::List recordsList(keys.size());
+    recordsList.names() = Rcpp::wrap(std::vector<std::string>(keys.begin(), keys.end()));
+    for (const std::string &k : keys)
+    {
+        std::vector<std::string> vals;
+        for (const auto &r : records)
+        {
+            const auto it = r.values.find(k);
+            vals.push_back((it == r.values.end()) ? "NA" : it->second);
+        }
+        recordsList[k] = vals;
+    }
+    
+    Rcpp::List specList(records.size());
+    specList.names() = recordsList["DB_ID"];
+    for (int i=0; i<specList.size(); ++i)
+    {
+        Rcpp::NumericMatrix nm(records[i].spectrum.mzs.size(), 2);
+        nm(Rcpp::_, 0) = Rcpp::NumericVector(records[i].spectrum.mzs.begin(), records[i].spectrum.mzs.end());
+        nm(Rcpp::_, 1) = Rcpp::NumericVector(records[i].spectrum.intensities.begin(),
+           records[i].spectrum.intensities.end());
+        Rcpp::colnames(nm) = Rcpp::CharacterVector({"mz", "intensity"});
+        specList[i] = nm;
+    }
+    
+    return Rcpp::List::create(Rcpp::Named("records") = Rcpp::DataFrame(recordsList),
+                              Rcpp::Named("spectra") = specList);
+}
+
+}
+
+
 // [[Rcpp::export]]
 Rcpp::List readMSP(Rcpp::CharacterVector file, Rcpp::LogicalVector pc)
 {
     const bool pComments = Rcpp::as<bool>(pc);
     std::ifstream fs;
-    std::vector<MSPRecord> records;
+    std::vector<MSLibRecord> records;
     std::vector<std::string> keys;
     
     auto addKey = [&keys](const std::string &k)
@@ -59,7 +95,7 @@ Rcpp::List readMSP(Rcpp::CharacterVector file, Rcpp::LogicalVector pc)
     if (fs.is_open())
     {
         std::string line;
-        MSPRecord curRec;
+        MSLibRecord curRec;
         
         while (std::getline(fs, line))
         {
@@ -90,20 +126,20 @@ Rcpp::List readMSP(Rcpp::CharacterVector file, Rcpp::LogicalVector pc)
                         const std::string com = curRec.values["Comments"];
                         std::string cv;
                         if (curRec.values.find("SMILES") == curRec.values.end() &&
-                            (parseComments(com, "SMILES", cv) || parseComments(com, "computed SMILES", cv)))
+                            (parseMSPComments(com, "SMILES", cv) || parseMSPComments(com, "computed SMILES", cv)))
                         {
                             addKey("SMILES");
                             curRec.values["SMILES"] = cv;
                         }
                         if (curRec.values.find("InChI") == curRec.values.end() &&
[…]
