### 2021-12-24 065b5e4
screenSuspects/unset: make sure that group qualities/scores are transferred, unset: copy() objects
