+++ R/feature_groups-screening.R
@@ -570,7 +570,8 @@ setMethod("screenSuspects", "featureGroups", function(fGroups, suspects, rtWindo
     ret <- featureGroupsScreening(screenInfo = scr, groups = copy(groupTable(fGroups)),
                                   analysisInfo = analysisInfo(fGroups), groupInfo = groupInfo(fGroups),
                                   features = getFeatures(fGroups), ftindex = copy(groupFeatIndex(fGroups)),
-                                  annotations = copy(annotations(fGroups)))
+                                  groupQualities = copy(groupQualities(fGroups)),
+                                  groupScores = copy(groupScores(fGroups)), annotations = copy(annotations(fGroups)))
     
     saveCacheData("screenSuspects", ret, hash)
     
