### 2021-12-24 d868dd9
small debug fix [skip appveyor]
