### 2021-12-24 0728b3c
initial WIP screenISTDs()
