+++ R/feature_groups.R
@@ -1093,6 +1093,36 @@ setMethod("selectIons", "featureGroups", function(fGroups, components, prefAdduc
     return(fGroups)
 })
 
+setMethod("screenISTDs", "featureGroups", function(fGroups, standards, refRGroups, rtWindow = 12, mzWindow = 0.005,
+                                                   skipInvalid = TRUE, adduct = NULL)
+{
+    checkmate::assertSubset(refRGroups, replicateGroups(fGroups), empty.ok = FALSE)
+    
+    anaInfo <- analysisInfo(fGroups)
+    if (is.null(anaInfo[["istd_conc"]]))
+        stop("No internal standard concentrations defined: no istd_conc column in analysis information", call. = FALSE)
+    if (any(is.na(anaInfo[anaInfo$group %in% refRGroups, "istd_conc"])))
+        stop("One or more reference internal standard concentrations are NA", call. = FALSE)
+        
+    # HACK: what we should do here for screening is exactly the same as screenSuspects(). So simply call that and use
+    # its output...
+    fGroupsScr <- screenSuspects(fGroups, suspects = standards, rtWindow = rtWindow, mzWindow = mzWindow,
+                                 skipInvalid = skipInvalid, adduct = adduct)
+    fGroups@iSTDs <- screenInfo(fGroupsScr)
+    origN <- uniqueN(fGroups@iSTDs$name)
+    
+    # only keep hits that are present in the analyses with non-NA conc
+    fGroupsWithISTD <- fGroups[, fGroups@iSTDs$group]
+    fGroupsWithISTD <- fGroupsWithISTD[!is.na(anaInfo$istd_conc)]
+    fGroupsWithISTD <- minAnalysesFilter(fGroupsWithISTD, relThreshold = 1, verbose = FALSE)
+    
+    fGroups@iSTDs <- fGroups@iSTDs[group %in% names(fGroupsWithISTD)]
+    
+    printf("Removed %d non-ubiquitous internal standards\n", origN - uniqueN(fGroups@iSTDs))
+    
+    return(fGroups)
+})
+
 #' Grouping of features
 #'
 #' Group equal features across analyses.
