### 2021-12-26 1f5bbf4
parse json metadata
