### 2021-12-27 09333c0
read spectra
