### 2021-12-27 06285d9
make loading of InChI/InChIKeys optional and also try to get them from meta data
