### 2021-12-27 29add3d
load annotations as vectors [skip ci]
