### 2021-12-28 41f28f6
handle parse errors
