### 2021-12-29 b118d07
read names and SPLASH
