### 2021-12-29 2c921dd
skip MoNA in-source MSP Synom markers
