### 2021-12-29 499d1b3
better (MoNA) consistency between MAP/json loading
