### 2021-12-29 04f7bfc
R interface to MoNA JSON loading [skip ci]
