+++ R/MSLibrary.R
@@ -12,6 +12,157 @@ makeDBIDsUnique <- function(records)
     return(records)
 }
 
+sanitizeMSLibrary <- function(lib, potAdducts, absMzDev, calcSPLASH)
+{
+    printf("Converting to tables\n")
+    lib$records <- as.data.table(lib$records)
+    # UNDONE: not for now, takes a lot of time for large amounts
+    # lib$spectra <- withProg(length(lib$spectra), FALSE, lapply(lib$spectra, function(s)
+    # {
+    #     ret <- as.data.table(s)
+    #     doProgress()
+    #     return(ret)
+    # }))
+    
+    # C++ code sets "NA" as string, convert to NA. Similarly, library may have 'n/a' markers...
+    for (j in seq_along(lib$records))
+        set(lib$records, which(lib$records[[j]] %chin% c("NA", "n/a", "N/A")), j, NA_character_)
+    
+    # Ensure case of column names used by patRoon are consistent
+    chCols <- c("Name", "SMILES", "InChI", "InChIKey", "Formula", "Precursor_type", "Ion_mode", "SPLASH")
+    numCols <- c("ExactMass", "PrecursorMZ")
+    allCols <- c(chCols, numCols)
+    change <- match(tolower(allCols), tolower(names(lib$records)), nomatch = integer())
+    setnames(lib$records, change, allCols)
+    
+    if (!all(c("Name", "DB_ID") %in% names(lib$records)))
+        stop("MSP file misses mandatory Name and/or DB# data.")
+    
+    lib$records <- makeDBIDsUnique(lib$records)
+    names(lib$spectra) <- lib$records$DB_ID
+    
+    # make sure columns at least exist, which makes future checks easier
+    for (col in allCols)
+    {
+        if (is.null(lib$records[[col]]))
+            lib$records[, (col) := if (col %in% chCols) NA_character_ else NA_real_]
+    }
+    
+    # cleanup character data: trim whitespace and convert empties to NA
+    lib$records[, (chCols) := lapply(.SD, function(x)
+    {
+        x <- trimws(x)
+        return(fifelse(nzchar(x), x, NA_character_))
+    }), .SDcols = chCols]
+    
+    # ensure mass data is numeric, but ignore conversion warnings
+    suppressWarnings({
+        lib$records[, (numCols) := lapply(.SD, as.numeric), .SDcols = numCols]
+        if (!is.null(lib$records[["MW"]]))
+            lib$records[, MW := as.numeric(MW)] # MW is not used by patRoon, but still make it numeric if present
+    })
+    
+    # UNDONE: this is quite slow, skip for now?
+    # printf("Sanitizing SMILES/InChI values... ")
+    # lib$records[!is.na(SMILES), SMILES := babelConvert(SMILES, "smi", "smi", FALSE)]
+    # lib$records[!is.na(InChI), InChI := babelConvert(InChI, "inchi", "inchi", FALSE)]
+    # printf("Done!\n")
+    
+    lib$records <- convertChemDataIfNeeded(lib$records, destFormat = "smi", destCol = "SMILES",
+                                           fromFormats = "inchi", fromCols = "InChI")
+    lib$records <- convertChemDataIfNeeded(lib$records, destFormat = "inchi", destCol = "InChI",
+                                           fromFormats = "smi", fromCols = "SMILES")
+    lib$records <- convertChemDataIfNeeded(lib$records, destFormat = "inchikey", destCol = "InChIKey",
+                                           fromFormats = c("smi", "inchi"), fromCols = c("SMILES", "InChI"))
[…]
