+++ R/compounds-library.R
@@ -29,18 +29,20 @@ unifyLibNames <- function(cTab)
 
 #' @export
 setMethod("generateCompoundsLibrary", "featureGroups", function(fGroups, MSPeakLists, MSLibrary, minSim = 0.75,
-                                                                absMzDev = 0.002, adduct = NULL, checkIons = "adduct",
+                                                                minAnnSim = minSim, absMzDev = 0.002, adduct = NULL,
+                                                                checkIons = "adduct",
                                                                 specSimParams = getDefSpecSimParams())
 {
     # UNDONE: cache
     # UNDONE: show mirror spectrum in report? Would need library data somehow
     # UNDONE: don't normalize scores (or already not done?)
     # UNDONE: separate specSimParams for lib? E.g. to assume that lib spectra are cleaner and don't need intensity cleaning
+    # UNDONE: work with IK1?
 
     ac <- checkmate::makeAssertCollection()
     checkmate::assertClass(MSPeakLists, "MSPeakLists", add = ac)
     checkmate::assertClass(MSLibrary, "MSLibrary", add = ac)
-    aapply(checkmate::assertNumber, . ~ minSim + absMzDev, lower = 0, finite = TRUE, fixed = list(add = ac))
+    aapply(checkmate::assertNumber, . ~ minSim + minAnnSim + absMzDev, lower = 0, finite = TRUE, fixed = list(add = ac))
     checkmate::assertChoice(checkIons, c("adduct", "polarity", "none"), add = ac)
     assertSpecSimParams(specSimParams, add = ac)
     checkmate::reportAssertions(ac)
@@ -58,6 +60,7 @@ setMethod("generateCompoundsLibrary", "featureGroups", function(fGroups, MSPeakL
     annTbl <- annotations(fGroups)
     libRecs <- records(MSLibrary)
     libSpecs <- spectra(MSLibrary)
+    libAnn <- annotations(MSLibrary)
     
     # UNDONE: ms level? or just a filter?
     # UNDONE: support entries without SMILES/InChI(keys)/Formulas?
@@ -116,7 +119,7 @@ setMethod("generateCompoundsLibrary", "featureGroups", function(fGroups, MSPeakL
 
         cTab <- unifyLibNames(cTab)
         cTab[, InChIKey1 := getIKBlock1(InChIKey)]
-        lspecs <- Map(libSpecs[cTab$identifier], cTab$ion_formula_mz, f = function(sp, pmz)
+        lspecs <- Map(libSpecs[cTab$identifier], cTab$ion_formula_mz, cTab$identifier, f = function(sp, pmz, lid)
         {
             # convert to MSPeakLists format
             ret <- as.data.table(sp)
@@ -124,6 +127,11 @@ setMethod("generateCompoundsLibrary", "featureGroups", function(fGroups, MSPeakL
             ret <- assignPrecursorToMSPeakList(ret, pmz)
             ret <- prepSpecSimilarityPL(ret, removePrecursor = specSimParams$removePrecursor,
                                         relMinIntensity = specSimParams$relMinIntensity, minPeaks = specSimParams$minPeaks)
+            
+            # add annotations, if any
+            if (!is.null(libAnn[[lid]]) && length(libAnn[[lid]]) > 0) # UNDONE: prune empty annotations while loading the library?
+                ret[, annotation := libAnn[[lid]][ID]]
+            
             return(ret)
         })
         lspecs <- pruneList(lspecs, checkZeroRows = TRUE)
@@ -136,6 +144,8 @@ setMethod("generateCompoundsLibrary", "featureGroups", function(fGroups, MSPeakL
         
         cTab[, score := sims[1, ]]
         
[…]
