### 2021-12-30 7aa7196
splitFormulaToList: support rcdk isotope format instead of Bruker format
