### 2021-12-30 9ec4716
proper super scripting of isotope numbers
