### 2021-12-30 fbcd0f1
add annotations to fragInfo
