### 2021-12-30 194e2a0
handle invalid formulas [skip ci]
