+++ R/utils-formulas.R
@@ -167,6 +167,13 @@ subscriptFormulaHTML <- function(formulas)
     return(ret)
 }
 
+verifyFormulas <- function(formulas)
+{
+    data("isotopes", package = "enviPat", envir = environment())
+    formulas[is.na(formulas)] <- "" # check_chemform() doesn't handle NAs
+    return(!enviPat::check_chemform(isotopes, formulas)$warning)
+}
+
 averageFormulas <- function(formulas)
 {
     fltab <- rbindlist(lapply(formulas, function(f) as.list(splitFormulaToList(f))), fill = TRUE)
