+++ R/compounds.R
@@ -66,7 +66,8 @@ setMethod("mergedConsensusNames", "compoundsConsensus", function(obj) obj@merged
 
 #' @describeIn compounds Returns default scorings that are excluded from normalization.
 #' @export
-setMethod("defaultExclNormScores", "compounds", function(obj) c("score", "individualMoNAScore", "annoTypeCount"))
+setMethod("defaultExclNormScores", "compounds", function(obj) c("score", "individualMoNAScore", "annoTypeCount",
+                                                                "libMatch"))
 
 setMethod("annScoreNames", "compounds", function(obj, onlyNums) compScoreNames(onlyNums))
 
