+++ R/generics.R
@@ -395,7 +395,7 @@ setGeneric("compoundViewer", function(fGroups, MSPeakLists, compounds) standardG
 setGeneric("reportCSV", function(fGroups, path = "report", reportFeatures = FALSE, formulas = NULL,
                                  formulasNormalizeScores = "max", formulasExclNormScores = NULL,
                                  compounds = NULL, compoundsNormalizeScores = "max",
-                                 compoundsExclNormScores = c("score", "individualMoNAScore", "annoTypeCount"),
+                                 compoundsExclNormScores = c("score", "individualMoNAScore", "annoTypeCount", "libMatch"),
                                  compsCluster = NULL, components = NULL,
                                  retMin = TRUE, clearPath = FALSE) standardGeneric("reportCSV"))
 setGeneric("reportPDF", function(fGroups, path = "report", reportFGroups = TRUE,
@@ -403,7 +403,7 @@ setGeneric("reportPDF", function(fGroups, path = "report", reportFGroups = TRUE,
                                  formulasNormalizeScores = "max", formulasExclNormScores = NULL,
                                  reportFormulaSpectra = TRUE,
                                  compounds = NULL, compoundsNormalizeScores = "max",
-                                 compoundsExclNormScores = c("score", "individualMoNAScore", "annoTypeCount"),
+                                 compoundsExclNormScores = c("score", "individualMoNAScore", "annoTypeCount", "libMatch"),
                                  compoundsOnlyUsedScorings = TRUE, compoundsTopMost = 5,
                                  compsCluster = NULL, components = NULL, MSPeakLists = NULL, retMin = TRUE,
                                  EICGrid = c(2, 1), EICRtWindow = 20, EICMzExpWindow = 0.001, EICTopMost = 1,
@@ -413,7 +413,7 @@ setGeneric("reportHTML", function(fGroups, path = "report", reportPlots = c("cho
                                   formulas = NULL, formulasTopMost = 5,
                                   formulasNormalizeScores = "max", formulasExclNormScores = NULL,
                                   compounds = NULL, compoundsNormalizeScores = "max",
-                                  compoundsExclNormScores = c("score", "individualMoNAScore", "annoTypeCount"),
+                                  compoundsExclNormScores = c("score", "individualMoNAScore", "annoTypeCount", "libMatch"),
                                   compoundsOnlyUsedScorings = TRUE, compoundsTopMost = 5, compsCluster = NULL,
                                   includeMFWebLinks = "compounds", components = NULL, interactiveHeat = FALSE,
                                   MSPeakLists = NULL, specSimParams = getDefSpecSimParams(), TPs = NULL,
