### 2021-12-31 347a728
filter() and delete() [skip ci]
