+++ R/mslibrary.R
@@ -24,6 +24,11 @@ sanitizeMSLibrary <- function(lib, potAdducts, absMzDev, calcSPLASH)
     #     return(ret)
     # }))
     
+    lib$spectra <- pruneList(lib$spectra, checkEmptyElements = TRUE)
+    lib$annotations <- pruneList(lib$annotations, checkEmptyElements = TRUE)
+    lib$annotations <- lib$annotations[names(lib$annotations) %chin% names(lib$spectra)]
+    lib$records <- lib$records[DB_ID %chin% names(lib$spectra)]
+    
     # C++ code sets "NA" as string, convert to NA. Similarly, library may have 'n/a' markers...
     for (j in seq_along(lib$records))
         set(lib$records, which(lib$records[[j]] %chin% c("NA", "n/a", "N/A")), j, NA_character_)
@@ -171,6 +176,13 @@ sanitizeMSLibrary <- function(lib, potAdducts, absMzDev, calcSPLASH)
 MSLibrary <- setClass("MSLibrary", slots = c(records = "data.table", spectra = "list", annotations = "list"),
                       contains = "workflowStep")
 
+setMethod("initialize", "MSLibrary", function(.Object, ...)
+{
+    .Object <- callNextMethod(.Object, ...)
+    .Object@annotations <- makeEmptyListNamed(.Object@annotations)
+    return(.Object)
+})
+
 #' @export
 setMethod("records", "MSLibrary", function(obj) obj@records)
 
@@ -201,10 +213,8 @@ setMethod("[", c("MSLibrary", "ANY", "missing", "missing"), function(x, i, ...)
     if (!missing(i))
     {
         i <- assertSubsetArgAndToChr(i, names(x))
-        x@records <- x@records[DB_ID %in% i]
-        x@spectra <- x@spectra[i]
+        x <- delete(x, setdiff(names(x), i))
     }
-    
     return(x)
 })
 
@@ -226,6 +236,157 @@ setMethod("as.data.table", "MSLibrary", function(x)
 {
     allSpecs <- rbindlist(spectra(x), idcol = "DB_ID")
     return(merge(records(x), allSpecs, by = "DB_ID"))
+})
+
+#' @export
+setMethod("delete", "MSLibrary", function(obj, i = NULL, j = NULL, ...)
+{
+    if (!is.function(i))
+        i <- assertDeleteArgAndToChr(i, names(obj))
+    checkmate::assert(
+        checkmate::checkFunction(j, null.ok = TRUE),
+        checkmate::checkIntegerish(j, any.missing = FALSE, null.ok = TRUE),
+        .var.name = "j"
+    )
+    if (is.function(i) && is.function(j))
+        stop("i and j cannot both be functions")
+
+    if (length(i) == 0 || (!is.null(j) && length(j) == 0))
+        return(obj) # nothing to remove...
+    
+    # i = vector remove specified records
+    # i = function: remove returned records
+    # j = vector/function: index of peaks to remove from spectra of i (all if i=NULL)
+    
+    if (is.null(j))
+    {
+        if (!is.function(i))
+            obj@records <- obj@records[!DB_ID %chin% i]
+        else
+        {
+            rm <- i(obj@records, ...)
+            if (is.logical(rm))
+                obj@records <- obj@records[!rm]
+            else if (is.character(rm))
+                obj@records <- obj@records[!rm %chin% DB_ID]
+            else
+                obj@records <- obj@records[setdiff(seq_len(nrow(obj@records)), rm)]
+        }
+        
+        obj@spectra <- obj@spectra[obj@records$DB_ID]
+        if (length(obj@annotations) > 0)
[…]
