### 2021-12-31 #4029 @whedon
[REVIEW]: patRoon 2.0: Improved non-target analysis workflows including automated transformation product screening
**Submitting author:** @rickhelmus (<a href="http://orcid.org/0000-0001-9401-3133">Rick Helmus</a>)
**Repository:** https://github.com/rickhelmus/patRoon
**Branch with paper.md** (empty if default branch): paper20
**Version:** v2.0.1
