### 2022-01-01 cb7e213
remove some old and unneeded code
