### 2022-01-06 rickhelmus/patRoon#40 @hechth
[JOSS Review] State of the field: Do the authors describe how this software compares to other commonly-used packages?
The paper lacks a section giving an overview of the field and the actually used methods (short description of BioTransformer). Currently, this is somewhat addressed with Table 1 and the Statement of Need, but it is not a clear review of the state of the art.

If there is a reason for merging these two sections, please let me know and we can discuss things here 😄.

 There exist multiple publications on the prediction of transformation products and reaction pathways ([10.1021/acs.chemrestox.0c00224](https://pubs.acs.org/doi/10.1021/acs.chemrestox.0c00224); [10.1021/ci200542m](https://pubs.acs.org/doi/10.1021/ci200542m); [10.1016/j.ymben.2021.11.009](https://doi.org/10.1016/j.ymben.2021.11.009);) - it would be great if the section could give a small overview and argue why BioTransformer was chosen in particular.
