### 2022-01-13 5215875
updated future R package reference and fixed several DOIs [skip ci]
