+++ paper.bib
@@ -85,7 +85,7 @@
   pages = {116409},
   year = {2021},
   issn = {0165-9936},
-  doi = {https://doi.org/10.1016/j.trac.2021.116409},
+  doi = {10.1016/j.trac.2021.116409},
   url = {https://www.sciencedirect.com/science/article/pii/S0165993621002326},
   author = {Dong Li and Wenqing Liang and Xiaoxia Feng and Ting Ruan and Guibin Jiang}
 }
@@ -97,7 +97,7 @@
   volume = {1},
   number = {8},
   pages = {e217},
-  doi = {https://doi.org/10.1002/cpz1.217},
+  doi = {doi.org/10.1002/cpz1.217},
   url = {https://currentprotocols.onlinelibrary.wiley.com/doi/abs/10.1002/cpz1.217},
   eprint = {https://currentprotocols.onlinelibrary.wiley.com/doi/pdf/10.1002/cpz1.217},
   year = {2021}
@@ -123,7 +123,7 @@
   pages = {106885},
   year = {2022},
   issn = {0160-4120},
-  doi = {https://doi.org/10.1016/j.envint.2021.106885},
+  doi = {10.1016/j.envint.2021.106885},
   url = {https://www.sciencedirect.com/science/article/pii/S0160412021005109},
   author = {Jessy Krier and Randolph R. Singh and Todor Kondić and Adelene Lai and Philippe Diderich and Jian Zhang and Paul A. Thiessen and Evan E. Bolton and Emma L. Schymanski},
 }
@@ -275,15 +275,14 @@
   journal = {Environmental Science {\&} Technology}
 }
 
-@Misc{Bengtsson2020,
+@article{Bengtsson2020,
   author = {Henrik Bengtsson},
-  title = {A Unifying Framework for Parallel and Distributed Processing in {R} using Futures},
-  year = {2020},
-  month = {aug},
-  eprint = {2008.00553},
-  archiveprefix = {arXiv},
-  primaryclass = {cs.DC},
-  url = {https://arxiv.org/abs/2008.00553},
+  title = {{A Unifying Framework for Parallel and Distributed Processing
+          in R using Futures}},
+  year = {2021},
+  journal = {{The R Journal}},
+  doi = {10.32614/RJ-2021-048},
+  url = {https://journal.r-project.org/archive/2021/RJ-2021-048/index.html}
 }
 
 @article{Weininger88,
