### 2022-01-13 42220c7
make MS file extension checks case insensitive (closes #34)
