### 2022-01-14 96d079e
fixed another DOI (issue #32)
