+++ paper.bib
@@ -97,7 +97,7 @@
   volume = {1},
   number = {8},
   pages = {e217},
-  doi = {doi.org/10.1002/cpz1.217},
+  doi = {10.1002/cpz1.217},
   url = {https://currentprotocols.onlinelibrary.wiley.com/doi/abs/10.1002/cpz1.217},
   eprint = {https://currentprotocols.onlinelibrary.wiley.com/doi/pdf/10.1002/cpz1.217},
   year = {2021}
