### 2022-01-14 0da9bdf
remove single page handbook (issue #39)
