+++ vignettes/handbook.Rmd
@@ -1,46 +1,0 @@
----
-title: "patRoon handbook"
-author: "Rick Helmus"
-date: "`r Sys.Date()`"  
-header-includes:
-- \usepackage{fvextra}
-- \DefineVerbatimEnvironment{Highlighting}{Verbatim}{breaklines,commandchars=\\\{\}}
-vignette: >
-  %\VignetteIndexEntry{Vignette Title}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r include=FALSE}
-vignDir <- normalizePath(".", winslash = "/")
-hbDir <- function(f)
-{
-    # HACK HACK HACK: currently we temporary re-name the markdown files to be
-    # underscore prefixed (see make_docs.R)
-    if (pkgdown::in_pkgdown())
-        f <- paste0("_", f)
-    return(file.path("handbook", f))
-}
-```
-
-```{r child=hbDir("setup.Rmd")}
-```
-
-```{r child=hbDir("intro.Rmd")}
-```
-
-```{r child=hbDir("installation.Rmd")}
-```
-
-```{r child=hbDir("concepts.Rmd")}
-```
-
-```{r child=hbDir("workflow_gen.Rmd")}
-```
-
-```{r child=hbDir("processing.Rmd")}
-```
-
-```{r child=hbDir("advanced.Rmd")}
-```
-
