### 2022-01-15 97e152f
convertToSuspects: better averaging defaults, optional collapsing
