### 2022-01-15 e6e8e9f
collapse results by IK1
