### 2022-01-15 fe6b086
cache MSLibrary loading
