### 2022-01-18 2c46ac4
newProject: don't specify (non-existing) MSPeakLists argument to reportCSV
