### 2022-01-18 f4fa0d3
verifyDataCentroided() utility: in preparation for #36
