### 2022-01-20 4f6c54b
verifyDataCentroided(): work with anaInfo and cache results (issue #36)
