### 2022-01-20 46f6c04
Verify if MS data is centroided for algorithms that need it (issue #36)
