+++ R/features-xcms.R
@@ -39,7 +39,7 @@ setMethod("delete", "featuresXCMS", function(obj, i = NULL, j = NULL, ...)
 findFeaturesXCMS <- function(analysisInfo, method = "centWave", ..., verbose = TRUE)
 {
     ac <- checkmate::makeAssertCollection()
-    analysisInfo <- assertAndPrepareAnaInfo(analysisInfo, c("mzXML", "mzML"), add = ac)
+    analysisInfo <- assertAndPrepareAnaInfo(analysisInfo, verifyCentroided = TRUE, c("mzXML", "mzML"), add = ac)
     checkmate::assertString(method, min.chars = 1, add = ac)
     checkmate::assertFlag(verbose, add = ac)
     checkmate::reportAssertions(ac)
@@ -86,7 +86,7 @@ importFeaturesXCMS <- function(xs, analysisInfo)
 {
     ac <- checkmate::makeAssertCollection()
     checkmate::assertClass(xs, "xcmsSet", add = ac)
-    analysisInfo <- assertAndPrepareAnaInfo(analysisInfo, c("mzXML", "mzML"), add = ac)
+    analysisInfo <- assertAndPrepareAnaInfo(analysisInfo, c("mzXML", "mzML"), verifyCentroided = TRUE, add = ac)
     checkmate::reportAssertions(ac)
 
     feat <- importXCMSPeaks(xcms::peaks(xs), analysisInfo)
