+++ R/features-envipick.R
@@ -21,7 +21,7 @@ setMethod("initialize", "featuresEnviPick",
 findFeaturesEnviPick <- function(analysisInfo, ..., parallel = TRUE, verbose = TRUE)
 {
     ac <- checkmate::makeAssertCollection()
-    analysisInfo <- assertAndPrepareAnaInfo(analysisInfo, "mzXML")
+    analysisInfo <- assertAndPrepareAnaInfo(analysisInfo, "mzXML", verifyCentroided = TRUE)
     aapply(checkmate::assertFlag, . ~ parallel + verbose, fixed = list(add = ac))
     checkmate::reportAssertions(ac)
 
@@ -83,7 +83,7 @@ findFeaturesEnviPick <- function(analysisInfo, ..., parallel = TRUE, verbose = T
 importFeaturesEnviMass <- function(analysisInfo, enviProjPath)
 {
     ac <- checkmate::makeAssertCollection()
-    analysisInfo <- assertAndPrepareAnaInfo(analysisInfo, "mzXML", add = ac)
+    analysisInfo <- assertAndPrepareAnaInfo(analysisInfo, "mzXML", verifyCentroided = TRUE, add = ac)
     checkmate::assertDirectoryExists(enviProjPath, "r", add = ac)
     checkmate::assertDirectoryExists(file.path(enviProjPath, "peaklist"), "r", .var.name = "enviProjPath", add = ac)
     checkmate::reportAssertions(ac)
