### 2022-01-20 e05da1e
verifyDataCentroided(): bit more info in the warnings emitted (issue #36)
