### 2022-01-20 5d111c2
further tweaks in warning output
