### 2022-01-20 4da9894
allow conversion of equal input/output formats
