### 2022-01-20 d88b8ce
workaround so that mzXML test data can be used by mzR (related to #36 and ProteoWizard/pwiz#1954)
