+++ tests/testthat/helper_utils.R
@@ -206,9 +206,14 @@ doDATests <- function() !is.null(getOption("patRoon.test.DAAnalyses"))
 
 makeMZXMLs <- function(anaInfo)
 {
+    # BUG/HACK: mzR cannot read mzXML files produced by OpenMS. The latter sets the .mzML source file as parentFile,
+    # which is currently not supported by pwiz (pwiz keeps the original instrument raw data format). However, pwiz does
+    # recognize mzXML as source format so simply convert the files twice...
     convertMSFiles(anaInfo = anaInfo, from = "mzML", outPath = getWorkPath(), to = "mzXML",
-                   algorithm = "openms")
+                   algorithm = "openms", overWrite = TRUE)
     anaInfo$path <- getWorkPath()
+    convertMSFiles(anaInfo = anaInfo, from = "mzXML", outPath = getWorkPath(), to = "mzXML",
+                   algorithm = "openms", overWrite = TRUE)
     return(anaInfo)
 }
 
