### 2022-01-20 1748ff8
add more notes on need for centroiding (issue #36)
