+++ man/MSPeakLists-generation.Rd
@@ -156,6 +156,11 @@ The \code{getDefAvgPListParams} is used to create a parameter list
   for peak list averaging (discussed below).
 }
 \note{
+In most cases it will be necessary to centroid your MS input files. The only exception is \command{Bruker},
+  however, you will still need centroided \file{mzXML}/\file{mzML} files for \emph{e.g.} plotting chromatograms. In
+  this case the centroided MS files should be stored in the same directory as the raw \command{Bruker} \file{.d}
+  files. The \code{\link{convertMSFiles}} function can be used to centroid data.
+
 \code{generateMSPeakListsDA} requires that the \option{Component}
   column is active (Method-->Parameters-->Layouts-->Mass List Layout) in
   order to add isotopologue information.
