+++ man/feature-finding.Rd
@@ -346,6 +346,11 @@ largely dependent upon applied analysis methodology and instrumentation.
 The file format of analyses for \code{findFeaturesXCMS} and \code{findFeaturesXCMS3} must be \code{mzML} or
   \code{mzXML}.
 
+In most cases it will be necessary to centroid your MS input files. The only exception is \command{Bruker},
+  however, you will still need centroided \file{mzXML}/\file{mzML} files for \emph{e.g.} plotting chromatograms. In
+  this case the centroided MS files should be stored in the same directory as the raw \command{Bruker} \file{.d}
+  files. The \code{\link{convertMSFiles}} function can be used to centroid data.
+
 \code{makeSet} Currently does not support making sets from \code{\link{featuresSet}} objects.
 
 \code{findFeaturesBruker} only works with Bruker data files (\code{.d}
