+++ man/feature-grouping.Rd
@@ -332,6 +332,11 @@ The \code{data.frame} method for \code{groupFeatures} is a special case that cur
   \code{\link{XCMSnExp}} object (from the \pkg{xcms} package).
 }
 \note{
+In most cases it will be necessary to centroid your MS input files. The only exception is \command{Bruker},
+  however, you will still need centroided \file{mzXML}/\file{mzML} files for \emph{e.g.} plotting chromatograms. In
+  this case the centroided MS files should be stored in the same directory as the raw \command{Bruker} \file{.d}
+  files. The \code{\link{convertMSFiles}} function can be used to centroid data.
+
 \code{makeSet} Currently does not support making sets from \code{\link{featureGroupsSet}} objects.
 }
 \section{Sets workflows}{
