### 2022-01-20 1b5e336
workaround for r-lib/pkgdown@2022
