+++ R/components-openms.R
@@ -42,9 +42,6 @@ setMethod("initialize", "componentsOpenMS",
 #' @templateVar what \code{generateComponentsOpenMS}
 #' @template uses-multiProc
 #'
-#' @section Parallelization: Note that for caching purposes, the analyses files must always exist on the local host
-#'   computer, even if it is not participating in computations.
-#'
 #' @references \insertRef{Bielow2010}{patRoon}
 #'
 #' @aliases generateComponentsOpenMS
