### 2022-01-21 9974669
split findFeatures() docs
