+++ man-roxygen/analysisInfo-arg.R
@@ -1,0 +1,1 @@
+#' @param analysisInfo A \code{data.frame} with \link[=analysis-information]{Analysis information}.
