+++ man-roxygen/centroid_note.R
@@ -1,0 +1,4 @@
+#' @note In most cases it will be necessary to centroid your MS input files. The only exception is \command{Bruker},
+#'   however, you will still need centroided \file{mzXML}/\file{mzML} files for \emph{e.g.} plotting chromatograms. In
+#'   this case the centroided MS files should be stored in the same directory as the raw \command{Bruker} \file{.d}
+#'   files. The \code{\link{convertMSFiles}} function can be used to centroid data.
