### 2022-01-21 32fcd7d
split findFeaturesOpenMS docs
