+++ R/features-openms.R
@@ -10,83 +10,76 @@ setMethod("initialize", "featuresOpenMS",
           function(.Object, ...) callNextMethod(.Object, algorithm = "openms", ...))
 
 
-#' @details \code{findFeaturesOpenMS} uses the
-#'   \href{https://abibuilder.informatik.uni-tuebingen.de/archive/openms/Documentation/release/latest/html/TOPP_FeatureFinderMetabo.html}{FeatureFinderMetabo}
-#'    TOPP tool (see \url{http://www.openms.de}).
+#' Find features using OpenMS
 #'
-#' @note The file format of analyses for \code{findFeaturesOpenMS} must be
-#'   \file{mzML}. This functionality has been tested with OpenMS version >= 2.0.
-#'   Please make sure it is installed and its binaries are added to the PATH
-#'   environment variable or the \code{patRoon.path.OpenMS} option is set.
+#' uses the
+#' \href{https://abibuilder.informatik.uni-tuebingen.de/archive/openms/Documentation/release/latest/html/TOPP_FeatureFinderMetabo.html}{FeatureFinderMetabo}
+#' TOPP tool (see \url{http://www.openms.de}) to find features.
 #'
-#' @param noiseThrInt Noise intensity threshold. Sets
-#'   \code{algorithm:common:noise_threshold_int} option.
-#' @param chromSNR Minimum S/N of a mass trace. Sets
-#'   \code{algorithm:common:chrom_peak_snr} option.
-#' @param chromFWHM Expected chromatographic peak width (in seconds). Sets
-#'   \code{algorithm:common:chrom_fwhm} option.
-#' @param mzPPM Allowed mass deviation (ppm) for trace detection. Sets
-#'   \code{algorithm:mtd:mass_error_ppm}.
-#' @param reEstimateMTSD If \code{TRUE} then enables dynamic re-estimation of
-#'   m/z variance during mass trace collection stage. Sets
-#'   \code{algorithm:mtd:reestimate_mt_sd}.
-#' @param traceTermCriterion,traceTermOutliers,minSampleRate Termination
-#'   criterion for the extension of mass traces. See
+#' @templateVar algo OpenMS
+#' @templateVar do automatically find features
+#' @templateVar generic findFeatures
+#' @templateVar algoParam openms
+#' @template algo_generator
+#'
+#' @details This functionality has been tested with OpenMS version >= 2.0. Please make sure it is installed and its
+#'   binaries are added to the PATH environment variable or the \code{patRoon.path.OpenMS} option is set.
+#'
+#'   The file format of analyses must be \file{mzML}.
+#'
+#' @template centroid_note_mandatory
+#'
+#' @inheritParams findFeatures
+#'
+#' @param noiseThrInt Noise intensity threshold. Sets \code{algorithm:common:noise_threshold_int} option.
+#' @param chromSNR Minimum S/N of a mass trace. Sets \code{algorithm:common:chrom_peak_snr} option.
+#' @param chromFWHM Expected chromatographic peak width (in seconds). Sets \code{algorithm:common:chrom_fwhm} option.
+#' @param mzPPM Allowed mass deviation (ppm) for trace detection. Sets \code{algorithm:mtd:mass_error_ppm}.
+#' @param reEstimateMTSD If \code{TRUE} then enables dynamic re-estimation of m/z variance during mass trace collection
+#'   stage. Sets \code{algorithm:mtd:reestimate_mt_sd}.
+#' @param traceTermCriterion,traceTermOutliers,minSampleRate Termination criterion for the extension of mass traces. See
 #'   \href{https://abibuilder.informatik.uni-tuebingen.de/archive/openms/Documentation/release/latest/html/TOPP_FeatureFinderMetabo.html}{FeatureFinderMetabo}.
[…]
