+++ man-roxygen/algo_generator.R
@@ -1,0 +1,4 @@
+#' @details This function uses algorithms from <%=algo%> to <%=do%>. This function is called when calling
+#'   \code{<%=generic%>} with \code{algorithm="<%=algoParam%>"}.
+#' 
+#' @seealso \code{\link{<%=generic%>}} for more details and other algorithms.
