+++ man-roxygen/centroid_note_mandatory.R
@@ -1,0 +1,2 @@
+#' @details The input MS data files need to be centroided. The \code{\link{convertMSFiles}} function can be used to
+#'   centroid data.
