+++ man-roxygen/parallelization-cache_input.R
@@ -1,0 +1,2 @@
+#' @section Parallelization: Note that for caching purposes, the analyses files must always exist on the local host
+#'   computer, even if it is not participating in computations.
