### 2022-01-21 55a1b92
split findFeaturesSAFD docs
