### 2022-01-21 f3955b4
more docs splitting for findFeatures functions [skip ci]
