### 2022-01-22 355dffd
babelConvert(): also allow appended formula calculation or just formula calculation
