### 2022-01-22 2f7d44f
replaced now obsolete convertToFormulaBabel() with babelConvert() and removed the function
