+++ tests/testthat/test-screening.R
@@ -13,7 +13,7 @@ if (testWithSets())
 }
 susps[, InChI := babelConvert(SMILES, "smi", "inchi")]
 susps[, neutralMass := getNeutralMassFromSMILES(SMILES)]
-susps[, formula := convertToFormulaBabel(SMILES, "smi")]
+susps[, formula := babelConvert(SMILES, "smi", "formula")]
 
 fGroups <- getTestFGroups(getTestAnaInfoAnn(), noiseThrInt = 1E4)
 fGroupsScr <- doScreen(fGroups, susps, onlyHits = TRUE)
