### 2022-01-23 8cd4797
WIP prepareChemTable() [skip ci]
