### 2022-01-24 a160e22
split importFeatures() ref docs
