+++ man-roxygen/algo_importer.R
@@ -1,0 +1,4 @@
+#' @details This function imports data from <%=algo%>. This function is called when calling \code{<%=generic%>} with
+#'   \code{type="<%=algoParam%>"}.
+#'
+#' @seealso \code{\link{<%=generic%>}} for more details and other algorithms.
