+++ R/features-envipick.R
@@ -87,12 +87,23 @@ findFeaturesEnviPick <- function(analysisInfo, ..., parallel = TRUE, verbose = T
     return(ret)
 }
 
-#' @details \code{importFeaturesEnviMass} imports features from a project
-#'   generated by the \pkg{enviMass} package. NOTE: this functionality has only
-#'   been tested with older versions of \pkg{enviMass}.
+#' Imports features from enviMass
+#'
+#' Imports features from a project generated by the \pkg{enviMass} package.
+#'
+#' @templateVar algo enviMass
+#' @templateVar generic importFeatures
+#' @templateVar algoParam envimass
+#' @template algo_importer
+#'
+#' @inheritParams importFeatures
 #'
 #' @param enviProjPath The path of the enviMass project.
-#' @rdname feature-finding
+#'
+#' @note This functionality has only been tested with older versions of \pkg{enviMass}.
+#' 
+#' @inherit importFeatures return
+#' 
 #' @export
 importFeaturesEnviMass <- function(analysisInfo, enviProjPath)
 {
