+++ R/features-kpic2.R
@@ -164,13 +164,23 @@ findfeaturesKPIC2 <- function(analysisInfo, kmeans = TRUE, level = 1000, ..., pa
     return(ret)
 }
 
-#' @details \code{importfeaturesKPIC2} converts features obtained with with the
-#'   \pkg{KPIC2} package to a new \code{\link{features}} object.
+#' Imports features from KPIC2
 #'
-#' @param picsList A \code{list} with a \code{pics} objects obtained with
-#'   \code{\link[KPIC]{getPIC}} or \code{\link[KPIC]{getPIC.kmeans}} for each analysis.
+#' Imports feature data generated by the \pkg{KPIC2} package.
 #'
-#' @rdname feature-finding
+#' @templateVar algo KPIC2
+#' @templateVar generic importFeatures
+#' @templateVar algoParam kpic2
+#' @template algo_importer
+#'
+#' @inheritParams importFeatures
+#'
+#' @param picsList A \code{list} with a \code{pics} objects obtained with \code{\link[KPIC]{getPIC}} or
+#'   \code{\link[KPIC]{getPIC.kmeans}} for each analysis.
+#'
+#' @inherit findfeaturesKPIC2 references
+#' @inherit importFeatures return
+#' 
 #' @export
 importfeaturesKPIC2 <- function(picsList, analysisInfo)
 {
