+++ R/features-xcms3.R
@@ -90,13 +90,22 @@ findFeaturesXCMS3 <- function(analysisInfo, param = xcms::CentWaveParam(), ..., 
     return(ret)
 }
 
-#' @details \code{importFeaturesXCMS3} converts features from an existing
-#'   \code{\link{XCMSnExp}} object (obtained with the \pkg{xcms} package)
-#'   to a new \code{\link{features}} object.
+#' Imports features from XCMS (new interface)
+#'
+#' Imports feature data generated from an existing \code{\link{XCMSnExp}} object generated by the \pkg{xcms} package.
+#'
+#' @templateVar algo XCMS3
+#' @templateVar generic importFeatures
+#' @templateVar algoParam xcms3
+#' @template algo_importer
+#'
+#' @inheritParams importFeatures
 #'
 #' @param xdata An \code{\link{XCMSnExp}} object.
+#' 
+#' @inherit findFeaturesXCMS3 references
+#' @inherit importFeatures return
 #'
-#' @rdname feature-finding
 #' @export
 importFeaturesXCMS3 <- function(xdata, analysisInfo)
 {
