### 2022-01-24 adfea6a
move docs of makeSet() methods to own page
