### 2022-01-24 e8494e3
split ref docs for groupFeatures()
