### 2022-01-24 eda032d
split docs for algorithm specific groupFeatures functions
