+++ man-roxygen/feat-arg.R
@@ -1,0 +1,1 @@
+#' @param feat The \code{\link{features}} object with the features to be grouped. 
