+++ man-roxygen/rtalign-arg.R
@@ -1,0 +1,1 @@
+#' @param rtalign Set to \code{TRUE} to enable retention time alignment.

[…]
