### 2022-01-24 eb2c4ca
split refs of all importFeatureGroups functions
