+++ R/main.R
@@ -843,12 +843,9 @@ NULL
 #'   for each set separately, and the screening results are combined afterwards. The \code{sets} column in the
 #'   \code{screenInfo} data marks in which sets the suspect hit was found.
 #'
-#'   \code{importFeatureGroupsBrukerTASQ} Is currently not supported for sets workflows.
-#'
-#' @note Both \code{screenSuspects} and \code{importFeatureGroupsBrukerTASQ} may use the suspect names to base file
-#'   names used for reporting, logging etc. Therefore, it is important that these are file-compatible names. For this
-#'   purpose, \code{screenSuspects} will automatically try to convert long, non-unique and/or otherwise incompatible
-#'   suspect names.
+#' @note Both \code{screenSuspects} may use the suspect names to base file names used for reporting, logging etc.
+#'   Therefore, it is important that these are file-compatible names. For this purpose, \code{screenSuspects} will
+#'   automatically try to convert long, non-unique and/or otherwise incompatible suspect names.
 #'
 #' @name suspect-screening
 NULL
