### 2022-01-24 226508e
babelConvert: handle empty input and handle SMILES output [skip ci]
