+++ man-roxygen/sets-loadrawdata-RTalign-note.R
@@ -1,2 +1,2 @@
 #' @section Sets workflows: \code{loadRawData} and arguments related to retention time alignment are currently not
-#'   supported for sets workflows.
+#'   supported for \link[=sets-workflow]{sets workflows}.
