+++ R/main.R
@@ -155,82 +155,6 @@ NULL
 #' @name analysis-information
 NULL
 
-#' Grouping of features
-#'
-#' Functions and classes for grouping of features across analyses.
-#'
-#' After \link[=findFeatures]{features have been found} the logical next step is to align and group them across
-#' analyses. This process is necessary to allow comparison of features between multiple analyses, which otherwise would
-#' be difficult due to small deviations in retention and mass data. Thus, algorithms of 'feature groupers' are used to
-#' collect features with similar retention and mass data. In addition, advanced retention time alignment algorithms
-#' exist to enhance grouping of features even with relative large retention time deviations (\emph{e.g.} possibly
-#' observed from analyses collected over a long period). Like \link{findFeatures}, various algorithms are supported
-#' which may have many parameters that can be fine-tuned. This fine-tuning is likely to be necessary, since optimal
-#' settings often depend on applied methodology and instrumentation.
-#'
-#' @param algorithm A \code{character} that specifies the algorithm to be used: either \code{"openms"}, \code{"xcms"},
-#'   \code{"xcms3"} or \code{"kpic2"} (\code{features method}), or \code{"sirius"} (\code{data.frame} method).
-#' @param obj Either a \code{\link{features}} object to be grouped, or a \code{data.frame} with
-#'   \link[=analysis-information]{analysis info} to be passed to \code{groupFeaturesSIRIUS} or a \code{featureGroups}
-#'   objects passed to \code{makeSet}.
-#' @param feat The \code{\link{features}} to be grouped. \code{importFeatureGroupsBrukerPA} and
-#'   \code{importFeatureGroupsEnviMass} only support features generated by \code{\link{findFeaturesBruker}} and
-#'   \code{\link{importFeaturesEnviMass}}, respectively.
-#' @param analysisInfo A \code{data.frame} with \link[=analysis-information]{analysis info}.
-#' @param rtalign Enable retention time alignment.
-#' @param \dots Any parameters to be passed to the selected grouping/importing algorithm.
-#'
-#'   For \code{makeSet}: further \code{\link{featureGroups}} objects that should be used for the
-#'   \link[=sets-workflow]{sets workflow}.
-#' @param groupArgs named \code{character vector} that may contain extra grouping parameters to be used by
-#'   \code{\link[xcms:group-methods]{xcms::group}} or \code{\link[KPIC:PICset.group]{KPIC::PICset.group}}.
-#' @param verbose if \code{FALSE} then no text output will be shown.
-#' @param type Which file type should be imported or exported: \code{"brukerpa"} (Bruker ProfileAnalysis),
-#'   \code{"brukertasq"} (Bruker TASQ), \code{envimass} (\pkg{enviMass}, only import) or \code{"mzmine"} (MZMine, only
-#'   export).
-#'
-#' @template loadrawdata-arg
-#'
-#' @template centroid_note
-#'
-#' @section Sets workflows: With a \link[=sets-workflow]{sets workflow} the features are first neutralized, \emph{i.e.}
-#'   their \emph{m/z} values is replaced by their neutral monoisotopic mass. The latter is calculated from adduct
-#'   annotations, specified either directly, or already present in the input objects. The features are then grouped by
-#'   any of the \code{groupFeatures} algorithms. The neutralization step ensures that features measured with \emph{e.g.}
-#'   different ionization polarities can be grouped since their neutral mass will be the same. The sets workflow is
-#'   initiated in either of the following two ways: \enumerate{
-#'
-#'   \item The \code{groupFeatures} method can be called with a \code{\link{featuresSet}} object (generated with the
-#'   \link[=makeSet,features-method]{features method} of \code{makeSet}).
-#'
[…]
