### 2022-01-26 cd9e314
split ref docs for generateFormulas functions [skip ci]
