+++ R/formulas-bruker.R
@@ -15,25 +15,44 @@ getEmptyDAFragInfo <- function() data.table(mz = numeric(), ion_formula = charac
                                             mSigma = numeric(), dbe = numeric(), score = numeric())
 
 
-#' @details \code{generateFormulasDA} uses Bruker DataAnalysis to generate chemical formulae. This method supports
-#'   scoring based on overlap between measured and theoretical isotopic patterns (both MS and MS/MS data) and the
-#'   presence of 'fitting' MS/MS fragments. The method will iterate through all features (or "Compounds" in DataAnalysis
-#'   terms) and call \command{SmartFormula} (and \command{SmartFormula3D} if MS/MS data is available) to generate all
-#'   formulae. Parameters affecting formula calculation have to be set in advance within the DataAnalysis method for
-#'   each analysis (\emph{e.g.} by \code{\link{setDAMethod}}). This method requires that features were obtained with
-#'   \code{\link{findFeaturesBruker}}. It is recommended, but not mandatory, that the \code{\link{MSPeakLists}} are also
-#'   generated by DataAnalysis.
+#' Generate formula with Bruker DataAnalysis
+#'
+#' Uses Bruker DataAnalysis to generate chemical formulae.
+#'
+#' @templateVar algo Bruker DataAnalysis
+#' @templateVar do generate formula candidates
+#' @templateVar generic generateFormulas
+#' @templateVar algoParam bruker
+#' @template algo_generator
+#'
+#' @details This method supports scoring based on overlap between measured and theoretical isotopic patterns (both MS
+#'   and MS/MS data) and the presence of 'fitting' MS/MS fragments. The method will iterate through all features (or
+#'   "Compounds" in DataAnalysis terms) and call \command{SmartFormula} (and \command{SmartFormula3D} if MS/MS data is
+#'   available) to generate all formulae. Parameters affecting formula calculation have to be set in advance within the
+#'   DataAnalysis method for each analysis (\emph{e.g.} by \code{\link{setDAMethod}}).
+#'
+#'   This method requires that features were obtained with \code{\link{findFeaturesBruker}}. It is recommended, but not
+#'   mandatory, that the \code{\link{MSPeakLists}} are also generated by DataAnalysis.
+#'
+#'   Calculation of formulae with DataAnalysis always occurs with the 'feature approach' (see \verb{Candidate assignment} in \code{\link{generateFormulas}}).
 #'
 #' @param precursorMzSearchWindow Search window for \emph{m/z} values (+/- the feature \emph{m/z}) used to find back
 #'   feature data of precursor/parent ions from MS/MS spectra (this data is not readily available from
 #'   \command{SmartFormula3D} results).
 #'
+#' @template adduct-arg
+#' @templateVar algo bruker
+#' @template form_algo-args
+#'
+#' @inheritParams generateFormulas
+#'
+#' @inherit generateFormulas return
+#'
 #' @template dasaveclose-args
 #'
 #' @template DA-restart-note
 #'
-#' @aliases generateFormulasDA
-#' @rdname formula-generation
+#' @name generateFormulasDA
 #' @export
 setMethod("generateFormulasDA", "featureGroups", function(fGroups, MSPeakLists, precursorMzSearchWindow = 0.002,
                                                           MSMode = "both", adduct = NULL, featThreshold = 0,
@@ -276,7 +295,8 @@ setMethod("generateFormulasDA", "featureGroups", function(fGroups, MSPeakLists, 
     return(formulas(groupAnnotations = groupFormulas, featureFormulas = fTable, algorithm = "bruker"))
 })
 
-#' @rdname formula-generation
+#' @template featAnnSets-gen_args
+#' @rdname generateFormulasDA
 #' @export
[…]
