### 2022-01-27 3274560
remove old formula generation ref page and update links to it
