### 2022-01-27 6459778
split generateCompounds ref docs
