+++ R/compounds-metfrag.R
@@ -22,7 +22,7 @@ NULL
 #' @templateVar class compoundsMF
 #' @template class-hierarchy
 #'
-#' @seealso \code{\link{compounds}} and \link{compound-generation}
+#' @seealso \code{\link{compounds}} and \code{\link{generateCompoundsMetFrag}}
 #'
 #' @references \insertRef{Ruttkies2016}{patRoon}
 #'
@@ -429,124 +429,127 @@ MFMPErrorHandler <- function(cmd, exitStatus, retries)
                    cmd$gName, exitStatus, cmd$logFile))
 }
 
-#' @details \code{generateCompoundsMetFrag} uses the \pkg{metfRag} package or
-#'   \command{MetFrag CL} for compound identification (see
-#'   \url{http://ipb-halle.github.io/MetFrag/}). Several online compound
-#'   databases such as \href{https://pubchem.ncbi.nlm.nih.gov/}{PubChem} and
-#'   \href{http://www.chemspider.com/}{ChemSpider} may be chosen for retrieval
-#'   of candidate structures. In addition, many options exist to score and
-#'   filter resulting data, and it is highly suggested to optimize these to
-#'   improve results. While MS/MS data is not mandatory, it will usually greatly
-#'   improve candidate scoring. The \command{MetFrag} options \code{PeakList},
-#'   \code{IonizedPrecursorMass} and \code{ExperimentalRetentionTimeValue} (in
-#'   minutes) fields are automatically set from feature data.
+#' Compound annotation with MetFrag
 #'
-#' @param method Which method should be used for MetFrag execution: \code{"CL"}
-#'   for \command{MetFragCL} and \code{"R"} for \command{MetFragR}. The former
-#'   might be faster.
-#' @param timeout Maximum time (in seconds) before a metFrag query for a feature
-#'   group is stopped. Also see \code{timeoutRetries} argument.
-#' @param timeoutRetries Maximum number of retries after reaching a timeout
-#'   before completely skipping the metFrag query for a feature group. Also see
-#'   \code{timeout} argument.
-#' @param dbRelMzDev Relative mass deviation (in ppm) for database search. Sets
-#'   the \option{DatabaseSearchRelativeMassDeviation} option.
-#' @param fragRelMzDev Relative mass deviation (in ppm) for fragment matching.
-#'   Sets the \option{FragmentPeakMatchRelativeMassDeviation} option.
-#' @param fragAbsMzDev Absolute mass deviation (in Da) for fragment matching.
-#'   Sets the \option{FragmentPeakMatchAbsoluteMassDeviation} option.
-#' @param database Compound database to use. Valid values are: \code{"pubchem"},
-#'   \code{"chemspider"}, \code{"for-ident"}, \code{"comptox"},
-#'   \code{"pubchemlite"}, \code{"kegg"}, \code{"sdf"}, \code{"psv"} and
-#'   \code{"csv"}. See section below for more information. Sets the
-#'   \code{MetFragDatabaseType} option.
-#' @param extendedPubChem If \code{database="pubchem"}: whether to use the
-#'   \emph{extended} database that includes information for compound scoring
-#'   (\emph{i.e.} number of patents/PubMed references). Note that downloading
-#'   candidates from this database might take extra time. Valid values are:
-#'   \code{FALSE} (never use it), \code{TRUE} (always use it) or \code{"auto"}
-#'   (default, use if specified scorings demand it).
-#' @param chemSpiderToken A character string with the
-#'   \href{http://www.chemspider.com/AboutServices.aspx}{ChemSpider security
-#'   token} that should be set when the ChemSpider database is used. Sets the
-#'   \option{ChemSpiderToken} option.
-#' @param scoreTypes A character vector defining the scoring types. See the
-#'   \verb{Scorings} section below for more information. Note that both generic
[…]
