### 2022-01-27 17eb7a0
split generateComponents ref docs
