+++ R/components-tps.R
@@ -291,7 +291,7 @@ doGenComponentsTPs <- function(fGroups, fGroupsTPs, ignoreParents, TPs, MSPeakLi
 #'
 #' @param x,obj A \code{componentsTPs} object.
 #'
-#' @seealso \code{\link{components}} for other relevant methods and \link{component-generation}
+#' @seealso \code{\link{components}} for other relevant methods and \code{\link{generateComponents}}
 #'
 #' @templateVar class componentsTPs
 #' @template class-hierarchy
@@ -349,7 +349,7 @@ setMethod("as.data.table", "componentsTPs", function(x)
 #'   parameters} for more details). Set to \code{NULL} to ignore.
 #' @param minFragMatches,minNLMatches Minimum number of parent/TP fragment and neutral loss matches, respectively. Set
 #'   to \code{NULL} to ignore. See the \verb{Linking parents and transformation products} section in
-#'   \link{component-generation} for more details.
+#'   \code{\link{generateComponentsTPs}} for more details.
 #' @param formulas A \code{\link{formulas}} object. The formula annotation data in this object is to verify if elemental
 #'   additions/subtractions from metabolic logic reactions are possible (hence, it only works with data from
 #'   \code{\link{generateTPsLogic}}). To verify elemental additions, only TPs with at least one candidate formula that
@@ -491,12 +491,22 @@ setMethod("plotGraph", "componentsTPs", function(obj, onlyLinked)
     makeGraph(obj, onlyLinked, titles)
 })
 
-#' @details \code{generateComponentsTPs} generates components by linking feature groups of transformation products and
-#'   their parents. Moreover, this method typically employs data from \link[=TP-generation]{generated transformation
-#'   products} to find parents and their TPs. However, this data is not necessary, and components can also be made based
-#'   on MS/MS similarity and/or other annotation similarities between the parent and its TPs. For more details see the
+#' Generate components of transformation products
+#'
+#' Generates components by linking feature groups of transformation products and their parents.
+#'
+#' @templateVar algo transformation product screening
+#' @templateVar do generate components
+#' @templateVar generic generateComponents
+#' @templateVar algoParam tp
+#' @template algo_generator
+#'
+#' @details This method typically employs data from \link[=TP-generation]{generated transformation products} to find
+#'   parents and their TPs. However, this data is not necessary, and components can also be made based on MS/MS
+#'   similarity and/or other annotation similarities between the parent and its TPs. For more details see the
 #'   \verb{Linking parents and transformation products} section below.
 #'
+#' @param fGroups The input \code{\link{featureGroups}} for componentization. See \code{fGroupsTPs}.
 #' @param fGroupsTPs A \code{\link{featureGroups}} object containing the feature groups that are expected to be
 #'   transformation products. If a distinction between parents and TPs is not yet known, \code{fGroupsTPs} should equal
 #'   the \code{fGroups} argument. Otherwise, \code{fGroups} should only contain the parent feature groups, and both
@@ -504,17 +514,24 @@ setMethod("plotGraph", "componentsTPs", function(obj, onlyLinked)
 #' @param ignoreParents If \code{TRUE} then feature groups present in both \code{fGroups} and \code{fGroupsTPs} are not
 #'   considered as TPs.
 #' @param TPs A \code{\link{transformationProducts}} object. Set to \code{NULL} to perform linking without this data.
-#' @param formulas,compounds A \code{\link{formulas}}/\code{\link{compounds}} object to calculate annotation
-#'   similarities between parents and TPs. If \code{NULL} then this data is not calculated. For more details see the
-#'   \verb{Linking parents and transformation products} section below.
[…]
