+++ man-roxygen/compon_gen-clust.R
@@ -1,0 +1,4 @@
+#' @param method Clustering method that should be applied (passed to
+#'   \code{\link[fastcluster:hclust]{fastcluster::hclust}}).
+#' 
+#' @references \addCitations{fastcluster}{1}

[…]
