+++ R/main.R
@@ -301,70 +301,6 @@ NULL
 #' @name feature-optimization
 NULL
 
-#' Generation of transformation products (TPs)
-#'
-#' Functionality to automatically obtain transformation products for a given list of parent compounds.
-#'
-#' Several algorithms are available that obtain TPs for a set of parent compounds. Depending on the algorithm, the
-#' parents can be a suspect list, all the candidates in a \code{\link{compounds}} object or the screening results or all
-#' feature groups from a \code{\link{featureGroups}} object. The resulting objects can (again depending on the used
-#' algorithm) then be used for conversion to suspect lists (\code{\link{convertToSuspects}}), a \command{MetFrag}
-#' database (\code{\link{convertToMFDB}}) or further processed to generate parent to TP links from features
-#' (\code{\link{generateComponentsTPs}}).
-#'
-#' @param parents The parents for which transformation products should be obtained. This can be (1) a suspect list (see
-#'   \link[=suspect-screening]{suspect screening} for more information), (2) the resulting output
-#'   \code{\link{screenSuspects}} or (3) a \code{\link{compounds}} annotation object. In the former two cases, the
-#'   suspect (hits) are used as parents, whereas in the latter case all candidates are used as parents.
-#'
-#'   For \code{generateTPsLibrary}: can be \code{NULL}, in this case TPs for all parents in the library are obtained.
-#' @param skipInvalid If set to \code{TRUE} then the parents will be skipped (with a warning) for which insufficient
-#'   information (\emph{e.g.} SMILES) is available.
-#'
-#' @template adduct-arg
-#'
-#' @section Custom TP libraries and transformations: The \code{TPLibrary} argument is used to specify a custom TP
-#'   library for \code{generateTPsLibrary}. This should be a \code{data.frame} where each row specifies a TP for a
-#'   parent, with the following columns: \itemize{
-#'
-#'   \item \code{parent_name} and \code{TP_name}: The name of the parent/TP.
-#'
-#'   \item \code{parent_SMILES} and \code{TP_SMILES} The \acronym{SMILES} of the parent structure.
-#'
-#'   \item \code{parent_LogP} and \code{TP_LogP} The \code{log P} values for the parent/TP. (\strong{optional})
-#'
-#'   \item \code{LogPDiff} The difference between parent and TP \code{Log P} values. Ignored if \emph{both}
-#'   \code{parent_LogP} and \code{TP_LogP} are specified. (\strong{optional})
-#'
-#'   }
-#'
-#'   Other columns are allowed, and will be included in the final object. Multiple TPs for a single parent are specified
-#'   by repeating the value within \code{parent_} columns.
-#'
-#'   Similarly, the \code{transformations} argument to \code{generateTPsLogic} is used to specify custom rules to
-#'   calculate transformation products. This should be a \code{data.frame} with the following columns: \itemize{
-#'
-#'   \item \code{transformation} The name of the chemical transformation
-#'
-#'   \item \code{add} The elements that are added by this reaction (\emph{e.g.} \code{"O"}).
-#'
-#'   \item \code{sub} The elements that are removed by this reaction (\emph{e.g.} \code{"H2O"}).
-#'
-#'   \item \code{retDir} The expected retention time direction relative to the parent (assuming a reversed phase like LC
-#'   separation). Valid values are: \samp{-1} (elutes before the parent), \samp{1} (elutes after the parent) or \samp{0}
-#'   (no significant change or unknown).
-#'
-#'   }
-#'
-#' @note When the \code{parents} argument is a \code{\link{compounds}} object, the candidate library \code{identifier}
-#'   is used in case the candidate has no defined \code{compoundName}.
-#'
[…]

[…]
