### 2022-01-28 90bceb9
small fix for casing of KPIC2 function names
