+++ NAMESPACE
@@ -22,12 +22,12 @@ export(featureQualityNames)
 export(findFeatures)
 export(findFeaturesBruker)
 export(findFeaturesEnviPick)
+export(findFeaturesKPIC2)
 export(findFeaturesOpenMS)
 export(findFeaturesSAFD)
 export(findFeaturesSIRIUS)
 export(findFeaturesXCMS)
 export(findFeaturesXCMS3)
-export(findfeaturesKPIC2)
 export(formulaScorings)
 export(formulaTable)
 export(genIDLevelRulesFile)
@@ -59,9 +59,9 @@ export(importFeatureGroupsXCMS)
 export(importFeatureGroupsXCMS3)
 export(importFeatures)
 export(importFeaturesEnviMass)
+export(importFeaturesKPIC2)
 export(importFeaturesXCMS)
 export(importFeaturesXCMS3)
-export(importfeaturesKPIC2)
 export(newProject)
 export(numericIDLevel)
 export(optimizeFeatureFinding)
