### 2022-01-28 2df376f
misc fixes from devtools::check_man()
