+++ R/utils-compounds.R
@@ -41,15 +41,18 @@ mergeFragInfo <- function(fiLeft, fiRight, leftName, rightName)
 #'
 #' Returns an overview of scorings may be applied to rank candidate compounds.
 #'
+#' @param algorithm The algorithm: \code{"metfrag"} or \code{"sirius"}. Set to \code{NULL} to return all scorings.
+#' @param database The database for which results should be returned (\emph{e.g.} \code{"pubchem"}). Set to \code{NULL}
+#'   to return all scorings.
 #' @param includeSuspectLists,onlyDefault,includeNoDB A logical specifying whether scoring terms related to suspect
 #'   lists, default scoring terms and non-database specific scoring terms should be included in the output,
 #'   respectively.
 #'
 #' @return A \code{data.frame} with information on which scoring terms are used, what their algorithm specific name is
 #'   and other information such as to which database they apply and short remarks.
-#' 
+#'
 #' @seealso generateCompounds
-#' 
+#'
 #' @export
 compoundScorings <- function(algorithm = NULL, database = NULL, includeSuspectLists = TRUE,
                              onlyDefault = FALSE, includeNoDB = TRUE)
