+++ R/formulas-bruker.R
@@ -34,11 +34,15 @@ getEmptyDAFragInfo <- function() data.table(mz = numeric(), ion_formula = charac
 #'   This method requires that features were obtained with \code{\link{findFeaturesBruker}}. It is recommended, but not
 #'   mandatory, that the \code{\link{MSPeakLists}} are also generated by DataAnalysis.
 #'
-#'   Calculation of formulae with DataAnalysis always occurs with the 'feature approach' (see \verb{Candidate assignment} in \code{\link{generateFormulas}}).
+#'   Calculation of formulae with DataAnalysis always occurs with the 'feature approach' (see \verb{Candidate
+#'   assignment} in \code{\link{generateFormulas}}).
 #'
 #' @param precursorMzSearchWindow Search window for \emph{m/z} values (+/- the feature \emph{m/z}) used to find back
 #'   feature data of precursor/parent ions from MS/MS spectra (this data is not readily available from
 #'   \command{SmartFormula3D} results).
+#' @param MSMode Whether formulae should be generated only from MS data (\code{"ms"}), MS/MS data (\code{"msms"}) or
+#'   both (\code{"both"}). Selecting "both" will calculate formulae from MS data and MS/MS data and combines the results
+#'   (duplicated formulae are removed). This is useful when poor MS/MS data would exclude proper candidates.
 #'
 #' @template adduct-arg
 #' @templateVar algo bruker
