### 2022-01-28 46f7cb1
fix doc page and aliasing for workflow data generation functions
