+++ man-roxygen/main-rd-method.R
@@ -1,0 +1,3 @@
+#' @name <%=what%>
+#' @aliases <%=what%>
+#' @aliases <%=what%>,<%= if (exists("cl")) cl else "featureGroups" %>-method
