### 2022-01-28 d52b626
Some more KPIC2 casing fixes
