### 2022-01-28 8a9ce3c
reorganization of feature group filtering and plotting methods
- The ref documentation of fGroup plotting methods is now moved to a separate page.
- All fGroup plotting methods are moved to feature_groups-plot.R
- getFCParams() is moved to utils-export.R and has a separate doc page
- The fGroups filtering methods for sets workflows were moved to feature_groups-filter.R
