+++ R/feature_groups.R
@@ -16,48 +16,21 @@ NULL
 #' @param fGroups,obj,x,object \code{featureGroups} object to be accessed.
 #' @param rGroups For \code{[}: An optional \code{character} vector: if specified only keep results for the given
 #'   replicate groups (equivalent to the \code{rGroups} argument to \code{\link[=filter,featureGroups-method]{filter}}).
-#'
-#'   For \code{getFCParams}: a \code{character} vector with the names of the two replicate groups to be compared.
-#' @param retMin Plot retention time in minutes (instead of seconds).
 #' @param \dots For the \code{"["} operator: ignored.
 #'
 #'   For \code{delete}: passed to the function specified as \code{j}.
 #'
-#'   For \code{getFCParams}: optional named arguments that override defaults.
-#'   
 #'   \setsPassedArgs1{featureGroups}
-#'
-#'   Otherwise passed to \code{\link[graphics]{plot}} (\code{plot} and \code{plotChroms}), \code{\link[graphics]{lines}}
-#'   (\code{plotInt}), \pkg{\link{VennDiagram}} plotting functions (\code{plotVenn}), \code{\link{chordDiagram}}
-#'   (\code{plotChord}) or \code{\link[UpSetR]{upset}} (\code{plotUpSet}).
 #' @param average If \code{TRUE} then data within replicate groups are averaged.
 #'
 #'   For \code{as.data.table}: if \code{features=TRUE} other feature properties are also averaged.
-#' @param averageFunc Function used for averaging.
-#'   
-#'   for \code{as.data.table}: only used when \code{average=TRUE} or \code{FCParams != NULL}.
+#' @param averageFunc Function used for averaging. Only used when \code{average=TRUE} or \code{FCParams != NULL}.
 #' @param areas If set to \code{TRUE} then areas are considered instead of peak intensities.
 #'
 #'   For \code{as.data.table}: ignored if \code{features=TRUE}, as areas of features are always reported.
-#' @param pch,type,lty Common plotting parameters passed to \emph{e.g.} \code{\link[graphics]{plot}}. For \code{plot}:
-#'   if \code{pch=NULL} then values are automatically assigned.
-#' @param col Colour(s) used. If \code{col=NULL} then colours are automatically generated.
 #' @param which A character vector with replicate groups used for comparison.
-#'
-#'   For plotting functions: set to \code{NULL} for all replicate groups.
-#'
-#'   For \code{plotVenn}: alternatively a named \code{list} containing elements of \code{character} vectors with
-#'   replicate groups to compare. For instance, \code{which=list(infl = c("influent-A", "influent-B"), effl =
-#'   c("effluent-A", "effluent-B"))}, will compare the features in replicate groups \samp{"influent-A/B"} against those
-#'   in \samp{"effluent-A/B"}. The names of the list are used for labelling in the plot, and will be made automatically
-#'   if not specified.
-#' @param colourBy Sets the automatic colour selection: \code{"none"} for a single colour or
-#'   \code{"rGroups"}/\code{"fGroups"} for a distinct colour per replicate/feature group.
-#' @param showLegend Plot a legend if \code{TRUE}.
-#' @param FCParams A parameter list to calculate Fold change data (see \code{getFCParams} and the
-#'   \verb{Fold change calculation} section).
-#'   
-#'   For \code{as.data.table}: set to \code{NULL} to not perform FC calculations.
+#' @param FCParams A parameter list to calculate Fold change data. See \code{getFCParams} for more details. Set to
+#'   \code{NULL} to not perform FC calculations.
 #'
 #' @templateVar seli analyses
 #' @templateVar selOrderi analyses()
@@ -89,32 +62,11 @@ NULL
[…]
