+++ DESCRIPTION
@@ -194,6 +194,7 @@ Collate:
     'feature_groups-optimize-openms.R'
     'feature_groups-optimize-xcms.R'
     'feature_groups-optimize-xcms3.R'
+    'feature_groups-plot.R'
     'feature_groups-sirius.R'
     'feature_groups-tasq.R'
     'feature_groups-xcms.R'

[…]
