### 2022-01-28 201a527
doc update (in preparation for #35)
