### 2022-01-31 dab8316
small doc fixes
