+++ R/formulas-genform.R
@@ -294,11 +294,11 @@ processGenFormResultFile <- function(file, isMSMS, adduct, topMost)
 #' @template adduct-arg
 #' @templateVar algo genform
 #' @template form_algo-args
-#' 
+#'
 #' @inheritParams generateFormulas
 #'
 #' @inherit generateFormulas return
-#' 
+#'
 #' @section GenForm options: Below is a list of options (generated by running \command{GenForm} without commandline
 #'   options) which can be set by the \code{extraOpts} parameter.
 #'
@@ -309,7 +309,7 @@ processGenFormResultFile <- function(file, isMSMS, adduct, topMost)
 #' @template uses-multiProc
 #'
 #' @section Parallelization: When \code{futures} are used for parallel processing (\code{patRoon.MP.method="future"}),
-#'   calculations with \command{GenForm} are done with batch mode is disabled (see \code{batchSize} argument), which
+#'   calculations with \command{GenForm} are done with batch mode disabled (see \code{batchSize} argument), which
 #'   generally limit overall performance.
 #'
 #' @note This function always sets the \option{exist} and \option{oei} \command{GenForm} command line options.
