### 2022-01-31 #4029 @hechth
Just as quick update - the review is in progress, @rickhelmus already addressed some of my comments, so things are going great! I think it will take a bit more time since I also have to still test some things with different files (which is currently being delayed due to some internal issues). I don't think this will go beyond the 6 weeks usually required for the review.

In case a decision would have to be made before the formal end of the review, I'd highly recommend accepting the submission as the software and the paper are already of more than sufficient quality. Nevertheless, the points mentioned in the issues would be great to address before the publication.
