### 2022-01-31 5df28aa
handle InChIKeys for output
