### 2022-02-01 #4029 @KristinaRiemer
Hi @hechth, thank you for the update on the progress of your review! I agree that you and @rickhelmus should carry through with addressing all of your comments before publication.

@wesleyburr have you had a chance to start your review yet? Please check in with your progress and let me know if anything is blocking you!
