### 2022-02-03 c7aee71
more improvements on case insensitive MS file extension checks (issue #43)
