### 2022-02-03 1a6fed7
verify if mz(X)ML input files exist
