### 2022-02-03 50a7954
improve docs for analysis information (issue #33)
