+++ R/main.R
@@ -116,41 +116,51 @@ NULL
 #'   
 "_PACKAGE"
 
-#' Analysis information
+#' Properties of sample analyses
 #'
-#' Required information for analyses that should be processed and utilities to
-#' automatically generate this information.
+#' Properties for the sample analyses used in the workflow and utilities to automatically generate this information.
 #'
-#' @details Several properties need to be known about analyses that should be
-#'   processed during various workflow steps such as
-#'   \link{findFeatures}, averaging intensities of feature
-#'   groups and blank subtraction. This information should be made available
-#'   with an 'analysis info' object, which is a \code{data.frame} containing the
-#'   following columns:
+#' In \pkg{patRoon} a \emph{sample analysis}, or simply \emph{analysis}, refers to a single MS analysis file (sometimes
+#' also called \emph{sample} or \emph{file}). The \emph{analysis information} summarizes several properties for the
+#' analyses, and is used in various steps throughout the workflow, such as \code{\link{findFeatures}}, averaging
+#' intensities of feature groups and blank subtraction. This information should be in a \code{data.frame}, with the
+#' following columns:
 #'
-#' @template analysisInfo-list
+#' \itemize{
 #'
-#' @details Most functionality requires the data files to be in either the
-#'   \file{.mzXML} or \file{.mzML} format. Functionality that utilizes Bruker
-#'   DataAnalysis (\emph{e.g.} \code{\link{findFeaturesBruker}}) may only work
-#'   with data files in the proprietary \file{.d} format. Therefore, when tools
-#'   with varying requirements are mixed (a common scenario), the data files
-#'   also need to be present in the required formats. To deal with this
-#'   situation the data files with varying formats should all be placed in the
-#'   same path that was used to specify the location of the analysis.
+#' \item \code{path} the full path to the directory of the analysis.
 #'
-#'   Whether analysis data files are present in multiple formats or not, each
-#'   analysis should only be entered \emph{once}. The \code{analysis} column is
-#'   used as a basename to automatically find back the data file with the
-#'   required format, hence, analysis names should be specified as the file name
-#'   without its file extension.
+#' \item \code{analysis} the file name \strong{without} extension. Must be \strong{unique}, even if the \code{path} is
+#' different.
 #'
-#'   The \code{group} column is \emph{mandatory} and needs to be filled in for
-#'   each analysis. The \code{blank} column should also be present, however, these
-#'   may contain empty character strings (\code{""}) for analyses where no blank
-#'   subtraction should occur. The \code{conc} column is only required when
-#'   obtaining regression information is desired with the
-#'   \code{\link[=as.data.table,featureGroups-method]{as.data.table}} method.
+#' \item \code{group} name of \emph{replicate group}. A replicate group is used to group analyses together that are
+#' replicates of each other. Thus, the \code{group} column for all  analyses considered to be belonging to the same
+#' replicate group should have an equal (but unique) value. Used for \emph{e.g.} averaging and
+#' \code{\link[=filter,featureGroups-method]{filter}}.
+#'
+#' \item \code{blank}: all analyses within this replicate group are used by the \code{featureGroups} method of
+#' \code{\link[=filter,featureGroups-method]{filter}} for blank subtraction. Multiple entries can be entered by
+#' separation with a comma.
+#'
+#' \item \code{conc} a numeric value specifying the 'concentration' of the analysis. This can be actually any kind of
+#' numeric value such as exposure time, dilution factor or anything else which may be used to form a linear
+#' relationship.
+#'
+#' }
+#'
+#' Most workflows steps work with \file{mzXML} and \file{mzML} file formats. However, some algorithms only support
[…]
