+++ NEWS.md
@@ -14,6 +14,7 @@
 - Fixed: the `findFeaturesKPIC2()` and `importFeaturesKPIC2()` now have correct casing (was lower case 'f')
 - Fixed: some function arguments for `convertMSFiles()` were not properly verified
 - More checks to verify if input mzML/mzXML data actually exists
+- Improved reference documentation for `analysis-information` (issue #33)
 
 
 # patRoon 2.0

[…]
