+++ paper.bib
@@ -205,6 +205,12 @@
   journal = {Nucleic Acids Research}
 }
 
+@article{wolfe2016,
+  title={Chemical transformation system: Cloud based cheminformatic services to support integrated environmental modeling},
+  author={Wolfe, Kurt and Pope, Nicholas and Parmar, Rajbir and Galvin, Mike and Stevens, Caroline and Weber, Eric and Flaishans, Jonathan and Purucker, Tom},
+  year={2016}
+}
+
 @article{Senan2019,
   doi = {10.1093/bioinformatics/btz207},
   url = {https://doi.org/10.1093/bioinformatics/btz207},
