### 2022-02-08 dc822a2
more changes for #40
