+++ vignettes/handbook/workflow_gen.Rmd
@@ -1,5 +1,9 @@
 # Generating workflow data
 
+## Introduction
+
+### Workflow functions
+
 Each step in the non-target workflow is performed by a function that performs the heavy lifting of a workflow step behind the scenes and finally return the results. An important goal of `patRoon` is to support multiple algorithms for each workflow step, hence, when such a function is called you have to specify which algorithm you want to use. The available algorithms and their characteristics will be discussed in the next sections. An overview of all functions involved in generating workflow data is shown in the table below.
 
 Workflow step         | Function                                          | Output S4 class
@@ -13,7 +17,9 @@ MS peak lists         | `generateMSPeakLists()`                           | `MSP
 Formula annotation    | `generateFormulas()`                              | `formulas`
 Compound annotation   | `generateCompounds()`                             | `compounds`
 
-All of these functions store their output in objects derived from so called S4 classes. Knowing the details about the S4 class system of `R` is generally not important when using `patRoon` (and well written resources are available if you want to know more). In brief, usage of this class system allows a general data format that is used irrespective of the algorithm that was used to generate the data. For instance, when features have been found by [OpenMS] or [XCMS] they both return the same data format.
+### Workflow output
+
+The output of each workflow step is stored in objects derived from so called S4 classes. Knowing the details about the S4 class system of `R` is generally not important when using `patRoon` (and well written resources are available if you want to know more). In brief, usage of this class system allows a general data format that is used irrespective of the algorithm that was used to generate the data. For instance, when features have been found by [OpenMS] or [XCMS] they both return the same data format.
 
 Another advantage of the S4 class system is the usage of so called _generic functions_. To put simply: a generic function performs a certain task for different types of data objects. A good example is the `plotSpectrum()` function which plots an (annotated) spectrum from data of MS peak lists or from formula or compound annotation:
 
@@ -26,7 +32,112 @@ plotSpectrum(formulas, ...) # plot annotated spectrum from formula annotation da
 plotSpectrum(compounds, ...) # likewise but for compound annotation.
 ```
 
-The next sections will further detail on how to actually perform the non-target workflow steps to generate data.
+### Overview of all functions and their output
+
+```{r workflowFuncsClasses,echo=FALSE,out.width="100%",fig.cap="**Workflow functions and output classes.**"}
+
+makeRefURL <- function(node, page, label, tt) glue::glue("'{ node }' [URL=\"https://rickhelmus.github.io/patRoon/reference/{ page }.html\" label=<<U>{ label }</U>> tooltip=\"{ tt }\"]", node = node, page = page, label = label, tt = tt)
+makeRefURLClass <- function(class) makeRefURL(class, paste0(class, "-class"), class, paste(class, "class"))
+makeRefURLFunc <- function(func, page = func) makeRefURL(func, page, paste0(func, "()"), paste0(func, "()"))
+makeRefURLPage <- function(node, page) makeRefURL(node, page, node, node)
+
+URLs <- paste0(makeRefURLClass(c("transformationProducts", "featureGroupsScreening", "features", "featureGroups",
+                                 "MSPeakLists", "formulas", "compounds", "components")),
+               makeRefURLFunc(c("findFeatures", "groupFeatures", "generateTPs", "generateMSPeakLists",
[…]
