### 2022-02-09 ec8e520
switch back to pkgdown 2, with workarounds (see r-lib/pkgdown#2022) [skip appveyor]
