### 2022-02-11 4322a84
remove refRGroups
