+++ R/feature_groups-screening-set.R
@@ -421,7 +421,7 @@ setMethod("screenSuspects", "featureGroupsSet", function(fGroups, suspects, rtWi
                                      analysisInfo = analysisInfo(fGroups), groupInfo = groupInfo(fGroups),
                                      features = getFeatures(fGroups), ftindex = copy(groupFeatIndex(fGroups)),
                                      groupQualities = copy(groupQualities(fGroups)),
-                                     groupScores = copy(groupScores(fGroups)), iSTDs = copy(internalStandards(fGroups)),
+                                     groupScores = copy(groupScores(fGroups)), ISTDs = copy(internalStandards(fGroups)),
                                      annotations = copy(annotations(fGroups))))
 })
 
