### 2022-02-11 27d23ef
WIP normalizeIntensities
