### 2022-02-11 d2fb500
ISTDAssingments slot and few fixes
