+++ R/feature_groups-screening.R
@@ -572,7 +572,7 @@ setMethod("screenSuspects", "featureGroups", function(fGroups, suspects, rtWindo
                                   features = getFeatures(fGroups), ftindex = copy(groupFeatIndex(fGroups)),
                                   groupQualities = copy(groupQualities(fGroups)),
                                   groupScores = copy(groupScores(fGroups)), ISTDs = copy(internalStandards(fGroups)),
-                                  annotations = copy(annotations(fGroups)))
+                                  ISTDAssignments = fGroups@ISTDAssignments, annotations = copy(annotations(fGroups)))
     
     saveCacheData("screenSuspects", ret, hash)
     
