### 2022-02-11 ff423dc
normalizeIntensities(): ISTD RT window and minimum amount of ISTDs
