### 2022-02-11 2c36f3d
normalizeIntensities(): ISTD m/z window
