### 2022-02-11 1ebe4d2
TIC normalization
