### 2022-02-11 476fd16
normalizeIntensities() default normFunc to max [skip ci]
