### 2022-02-14 #4029 @hechth
There is only a single issue which would be good to address before release (https://github.com/rickhelmus/patRoon/issues/40), otherwise things are fine from my side - there are some bugs in the packages which are bundled within patRoon, but this is somewhat unavoidable and not the main responsibility of the author.
