### 2022-02-18 0e0eb45
merge screenISTDs() and normalizeIntensities()
