+++ R/generics.R
@@ -20,8 +20,8 @@ setGeneric("groupScores", function(fGroups) standardGeneric("groupScores"))
 setGeneric("internalStandards", function(fGroups) standardGeneric("internalStandards"))
 setGeneric("screenISTDs", function(fGroups, standards, rtWindow = 12, mzWindow = 0.005,
                                    skipInvalid = TRUE, ...) standardGeneric("screenISTDs"))
-setGeneric("normalizeIntensities", function(fGroups, method, ISTDRTWindow = 120, ISTDMZWindow = 300, minISTDs = 3,
-                                            normFunc = max) standardGeneric("normalizeIntensities"))
+setGeneric("normalizeIntensities", function(fGroups, method, normFunc = max, standards = NULL, ISTDRTWindow = 120,
+                                            ISTDMZWindow = 300, minISTDs = 3, ...) standardGeneric("normalizeIntensities"))
 
 setGeneric("getPICSet", function(obj, ...) standardGeneric("getPICSet"))
 
