### 2022-02-18 08dcc79
normalizeIntensities: method --> featNorm
