### 2022-02-18 b9355de
featNorm == "conc" to just normalize by istd_conc
