### 2022-02-18 f8f7700
remove now unneeded screenISTDs()
