+++ R/generics.R
@@ -18,8 +18,6 @@ setGeneric("replicateGroupSubtract", function(fGroups, rGroups, threshold = 0) s
 setGeneric("groupQualities", function(fGroups) standardGeneric("groupQualities"))
 setGeneric("groupScores", function(fGroups) standardGeneric("groupScores"))
 setGeneric("internalStandards", function(fGroups) standardGeneric("internalStandards"))
-setGeneric("screenISTDs", function(fGroups, standards, rtWindow = 12, mzWindow = 0.005,
-                                   skipInvalid = TRUE, ...) standardGeneric("screenISTDs"))
 setGeneric("normalizeIntensities", function(fGroups, featNorm = "none", normFunc = max, standards = NULL,
                                             ISTDRTWindow = 120, ISTDMZWindow = 300, minISTDs = 3,
                                             ...) standardGeneric("normalizeIntensities"))
