### 2022-02-18 8ec4177
make istd_conc optional if featNorm != "istd" | "conc"
