### 2022-02-18 59a4d65
normalizeIntensities: group normalization
