### 2022-02-20 0bfd4a6
improve cleaning DB formulas
