### 2022-02-20 fe56f04
babelConvert: fixup D elements
