### 2022-02-20 111d4b3
babelConvert: small fix for handling single output line
