### 2022-02-20 64f04a5
more prepareChemTable WIP [skip ci]
