### 2022-02-21 2427cc2
initial WIP for CompCreate.jl
