+++ NAMESPACE
@@ -85,6 +85,7 @@ exportClasses(components)
 exportClasses(componentsCamera)
 exportClasses(componentsCliqueMS)
 exportClasses(componentsClust)
+exportClasses(componentsCompCreate)
 exportClasses(componentsFeatures)
 exportClasses(componentsIntClust)
 exportClasses(componentsNT)
@@ -180,6 +181,7 @@ exportMethods(findFGroup)
 exportMethods(generateComponents)
 exportMethods(generateComponentsCAMERA)
 exportMethods(generateComponentsCliqueMS)
+exportMethods(generateComponentsCompCreate)
 exportMethods(generateComponentsIntClust)
 exportMethods(generateComponentsNontarget)
 exportMethods(generateComponentsOpenMS)
