### 2022-02-22 95852e6
Fix: update annotations slot of fGroups in delete()
