+++ TODO.md
@@ -1,7 +1,5 @@
 # Priority
 
-- Fix: update annotations slot of fGroups in delete()
-
 # Lower priority
 
 
