### 2022-02-22 d1a57c4
groupTable()/averageGroups(): support normalized data
