+++ R/generics.R
@@ -7,7 +7,8 @@ setGeneric("groupTable", function(object, ...) standardGeneric("groupTable"))
 setGeneric("groupFeatIndex", function(fGroups) standardGeneric("groupFeatIndex"))
 setGeneric("groupInfo", function(fGroups) standardGeneric("groupInfo"))
 setGeneric("removeEmptyAnalyses", function(fGroups) standardGeneric("removeEmptyAnalyses"))
-setGeneric("averageGroups", function(fGroups, areas = FALSE, func = mean) standardGeneric("averageGroups"))
+setGeneric("averageGroups", function(fGroups, areas = FALSE, normalized = FALSE,
+                                     func = mean) standardGeneric("averageGroups"))
 setGeneric("export", function(fGroups, type, out) standardGeneric("export"))
 setGeneric("unique", function(x, incomparables = FALSE, ...) standardGeneric("unique"))
 setGeneric("overlap", function(fGroups, which, exclusive = FALSE, ...) standardGeneric("overlap"))
