### 2022-02-22 00527b2
as.data.table(): use normalized feature data instead of calculating with normFunc
