### 2022-02-22 26f5f5b
as.data.table: add assigned ISTDs
