+++ R/feature_groups.R
@@ -749,6 +749,9 @@ setMethod("as.data.table", "featureGroups", function(x, average = FALSE, areas =
     if (nrow(ret) > 0 && nrow(annTable) > 0)
         ret <- merge(ret, annTable, sort = FALSE)
     
+    if (nrow(ret) > 0 && length(x@ISTDAssignments) > 0)
+        ret[, ISTD_assigned := sapply(x@ISTDAssignments[group], function(ia) paste0(ia, collapse = ","))]
+    
     return(ret[])
 })
 
