### 2022-02-22 9da8652
ensure that ISTDs have numeric relative intensities/areas
