### 2022-02-22 459f7ba
normalization for sets
