### 2022-02-22 da58f7d
delete(): also update values of ISTDAssignments
