### 2022-02-22 c8aec8f
delete(): only update entries for ISTDAssignments - not values
