### 2022-02-22 b05da5b
filter to remove ISTDs
