### 2022-02-23 rickhelmus/patRoon#40 @rickhelmus
Thank you for your suggestions.

While many tools are available, we specifically selected BioTransformer since (1) it allows prediction of TPs formed by pathways occurring in the environment, (2) is freely available and open-source and (3) allows good integration in automated workflows for batch predictions. This is now highlighted in the main text and the caption for Table 1. Furthermore, we clarified that future extension with other prediction tools is possible. Finally, we made the following improvements to Table 1:
-	The algorithms are highlighted if they are implemented or interfaced in patRoon
-	CTS was added as example TP prediction tool
-	Suspect screening of predicted or known TPs is now split over two rows to improve readability
-	Several minor textual improvements

I just triggered a new PDF build: [link](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04029/joss.04029/10.21105.joss.04029.pdf). (unfortuantely, there is now a page break which I cannot reproduce locally, I guess that's something to fix for later...).
