### 2022-02-23 034af4a
optimized group normalization
