### 2022-02-23 b92d769
auto normalization
