### 2022-02-23 6f59bca
small fixes for spectrumSimilarity()
