### 2022-02-23 e898b80
cleanup ISTD_assigned column
