### 2022-02-23 be09529
small unset() fix
