+++ R/feature_groups-set.R
@@ -487,5 +487,5 @@ setMethod("unset", "featureGroupsSet", function(obj, set)
                               features = unset(getFeatures(obj), set), ftindex = copy(groupFeatIndex(obj)),
                               groupQualities = copy(groupQualities(obj)), groupScores = copy(groupScores(obj)),
                               annotations = ann, ISTDs = copy(internalStandards(obj)),
-                              ISTDAssignments = obj@ISTDAssignments, algorithm = paste0(algorithm(obj), "_unset")))
+                              ISTDAssignments = obj@ISTDAssignments[[set]], algorithm = paste0(algorithm(obj), "_unset")))
 })
