### 2022-02-23 94270dc
delete()/unset(): update ISTDs slot
