### 2022-02-24 381bd4e
small fixes for normalizeIntensities()
