+++ R/feature_groups-sirius.R
@@ -93,8 +93,9 @@ setMethod("initialize", "featureGroupsSIRIUS",
 #'
 #' @details Finding and grouping features is done by running the \command{lcms-align} command on every analyses at once.
 #'   For this reason, grouping feature data from other algorithms than \command{SIRIUS} is not supported.
-#'   
-#'   The MS files should be in the \file{mzML} or \file{mzXML} format.
+#'
+#'   The MS files should be in the \file{mzML} or \file{mzXML} format. Furthermore, this algorithms requires the
+#'   presence of (data-dependent) MS/MS data.
 #'
 #' @template centroid_note_mandatory
 #'
