### 2022-02-24 5a21356
accessor method for ISTDAssignments slot
