+++ R/generics.R
@@ -19,6 +19,7 @@ setGeneric("replicateGroupSubtract", function(fGroups, rGroups, threshold = 0) s
 setGeneric("groupQualities", function(fGroups) standardGeneric("groupQualities"))
 setGeneric("groupScores", function(fGroups) standardGeneric("groupScores"))
 setGeneric("internalStandards", function(fGroups) standardGeneric("internalStandards"))
+setGeneric("internalStandardAssignments", function(fGroups, ...) standardGeneric("internalStandardAssignments"))
 setGeneric("normalizeIntensities", function(fGroups, featNorm = "none", groupNorm = FALSE, normFunc = max,
                                             standards = NULL, ISTDRTWindow = 120, ISTDMZWindow = 300, minISTDs = 3,
                                             ...) standardGeneric("normalizeIntensities"))
