+++ R/feature_groups-screening-set.R
@@ -388,7 +388,7 @@ setMethod("screenSuspects", "featureGroupsSet", function(fGroups, suspects, rtWi
                                      features = getFeatures(fGroups), ftindex = copy(groupFeatIndex(fGroups)),
                                      groupQualities = copy(groupQualities(fGroups)),
                                      groupScores = copy(groupScores(fGroups)), ISTDs = copy(internalStandards(fGroups)),
-                                     ISTDAssignments = fGroups@ISTDAssignments,
+                                     ISTDAssignments = internalStandardAssignments(fGroups),
                                      annotations = copy(annotations(fGroups))))
 })
 
