+++ NAMESPACE
@@ -218,6 +218,8 @@ exportMethods(groupQualities)
 exportMethods(groupScores)
 exportMethods(groupTable)
 exportMethods(identifiers)
+exportMethods(internalStandardAssignments)
+exportMethods(internalStandards)
 exportMethods(length)
 exportMethods(lengths)
 exportMethods(makeHCluster)
