### 2022-02-24 d819267
visualize ISTD assignments
