+++ TODO.md
@@ -38,6 +38,9 @@
     - split IS assignment in different function anyway?
     - as.data.table(): split ISTD_assignment column per set
     - find another way to assign close/far ISTDs: if there are multiple close ones available, it makes more sense to not consider those that are a bit far away.
+    - still subset ISTDAssignments? Needed for plotGraph
+        - or just verify if fGroups still exist in plotGraph?
+        - or keep as is? only affects tooltip
 
 
 ## Docs
