### 2022-03-01 99d5a59
removed onlyLinked argument from plotGraph generic
