### 2022-03-01 d82f877
plotGraph(): replaced onlyLinked argument with onlyPresent, which toggles visibility of removed internal standards
