### 2022-03-01 d66f4a4
as.data.table(): split ISTD_assignment column per set
