### 2022-03-02 28eafa3
use plotGraph for IS with reportHTML()
