+++ TODO.md
@@ -37,8 +37,6 @@
     - maybe: allow usage of normalized intensities with filter()?
     - split IS assignment in different function anyway?
     - find another way to assign close/far ISTDs: if there are multiple close ones available, it makes more sense to not consider those that are a bit far away.
-    - more reporting of ISTD/normalization related things?
-        - eg plotGraph()?
 
 
 ## Docs
@@ -52,6 +50,7 @@
 
 - Added slots to fGroups --> cache should be cleared
 - Removed onlyLinked argument from plotGraph generic (not components methods)
+- reportHTML: features menu with submenus
 
 
 # Lower priority
