### 2022-03-03 5379fbd
normalizeIntensities() --> normInts()
