+++ R/utils-features.R
@@ -122,6 +122,6 @@ maybeAutoNormalizeFGroups <- function(fGroups)
 {
     if (length(fGroups) == 0 || !is.null(featureTable(fGroups)[[1]][["intensity_rel"]]))
         return(fGroups) # no features or already normalized
-    printf("Automatically normalizing feature groups, see ?normalizeIntensities() for more options.\n")
-    return(normalizeIntensities(fGroups, featNorm = "none", groupNorm = TRUE))
+    printf("Automatically normalizing feature groups, see ?normInts for more options.\n")
+    return(normInts(fGroups, featNorm = "none", groupNorm = TRUE))
 }
