+++ R/generics.R
@@ -20,9 +20,8 @@ setGeneric("groupQualities", function(fGroups) standardGeneric("groupQualities")
 setGeneric("groupScores", function(fGroups) standardGeneric("groupScores"))
 setGeneric("internalStandards", function(fGroups) standardGeneric("internalStandards"))
 setGeneric("internalStandardAssignments", function(fGroups, ...) standardGeneric("internalStandardAssignments"))
-setGeneric("normalizeIntensities", function(fGroups, featNorm = "none", groupNorm = FALSE, normFunc = max,
-                                            standards = NULL, ISTDRTWindow = 120, ISTDMZWindow = 300, minISTDs = 3,
-                                            ...) standardGeneric("normalizeIntensities"))
+setGeneric("normInts", function(fGroups, featNorm = "none", groupNorm = FALSE, normFunc = max, standards = NULL,
+                                ISTDRTWindow = 120, ISTDMZWindow = 300, minISTDs = 3, ...) standardGeneric("normInts"))
 
 setGeneric("getPICSet", function(obj, ...) standardGeneric("getPICSet"))
 
