+++ TODO.md
@@ -33,7 +33,6 @@
 ## Features
 
 - Normalization
-    - rename normalizeIntensities()? normInts()? normalizeInts()?
     - maybe: allow usage of normalized intensities with filter()?
     - split IS assignment in different function anyway?
     - find another way to assign close/far ISTDs: if there are multiple close ones available, it makes more sense to not consider those that are a bit far away.
