### 2022-03-03 8cbe07f
convertMFDB() now always collapses duplicates, not just for BT
