### 2022-03-03 9d0b3d4
more WIP for TP structure base class: library/BT now subclass from it, constructor now handles parent similarity calculation and removed old methods
