+++ R/TP-biotransformer.R
@@ -1,5 +1,5 @@
 #' @include main.R
-#' @include TP.R
+#' @include TP-structure.R
 NULL
 
 #' Class to store transformation products (TPs) predicted by BioTransformer
@@ -21,13 +21,13 @@ NULL
 #' @template class-hierarchy
 #'
 #' @export
-transformationProductsBT <- setClass("transformationProductsBT", contains = "transformationProducts")
+transformationProductsBT <- setClass("transformationProductsBT", contains = "transformationProductsStructure")
 
 setMethod("initialize", "transformationProductsBT",
           function(.Object, ...) callNextMethod(.Object, algorithm = "biotransformer", ...))
 
 
-getBaseBTCmd <- function(parent, SMILES, type, steps, fpType, fpSimMethod, extraOpts, baseHash)
+getBaseBTCmd <- function(parent, SMILES, type, steps, extraOpts, baseHash)
 {
     mainArgs <- c("-b", type,
                   "-k", "pred",
@@ -36,8 +36,7 @@ getBaseBTCmd <- function(parent, SMILES, type, steps, fpType, fpSimMethod, extra
                   extraOpts)
     
     return(list(command = "java", args = mainArgs, logFile = paste0("biotr-", parent, ".txt"), parent = parent,
-                SMILES = SMILES, fpType = fpType, fpSimMethod = fpSimMethod,
-                hash = makeHash(parent, SMILES, baseHash)))
+                SMILES = SMILES, hash = makeHash(parent, SMILES, baseHash)))
 }
 
 collapseBTResults <- function(prod)
@@ -92,9 +91,6 @@ BTMPFinishHandler <- function(cmd)
     ret[, name := paste0(cmd$parent, "-TP", seq_len(nrow(ret)))]
     
     ret[, retDir := fifelse(ALogP < parent_ALogP, -1, 1)]
-    
-    ret[, similarity := mapply(parent_SMILES, SMILES, FUN = patRoon:::distSMILES,
-                               MoreArgs = list(fpType = cmd$fpType, fpSimMethod = cmd$fpSimMethod))][]
     
     return(ret)
 }
@@ -154,7 +150,8 @@ BTMPPrepareHandler <- function(cmd)
 #'
 #' @export
 generateTPsBioTransformer <- function(parents, type = "env", steps = 2, extraOpts = NULL,
-                                      skipInvalid = TRUE, fpType = "extended", fpSimMethod = "tanimoto", MP = FALSE)
+                                      skipInvalid = TRUE, calcSims = FALSE, fpType = "extended",
+                                      fpSimMethod = "tanimoto", MP = FALSE)
 {
     checkmate::assert(
         checkmate::checkClass(parents, "data.frame"),
@@ -170,9 +167,8 @@ generateTPsBioTransformer <- function(parents, type = "env", steps = 2, extraOpt
     checkmate::assertChoice(type, c("ecbased", "cyp450", "phaseII", "hgut", "superbio", "allHuman", "env"), add = ac)
     checkmate::assertCount(steps, positive = TRUE, add = ac)
     checkmate::assertCharacter(extraOpts, null.ok = TRUE, add = ac)
-    checkmate::assertFlag(skipInvalid, add = ac)
+    aapply(checkmate::assertFlag, . ~ skipInvalid + calcSims + MP, fixed = list(add = ac))
     aapply(checkmate::assertString, . ~ fpType + fpSimMethod, min.chars = 1, fixed = list(add = ac))
-    checkmate::assertFlag(MP, add = ac)
     checkmate::reportAssertions(ac)
 
     parents <- getTPParents(parents, skipInvalid)
[…]

[…]
