### 2022-03-04 40e96ac
make nontarget an optional dependency and install from GitHub (issue #48)
