### 2022-03-04 ab6ce01
also take care of nontargetData installation (issue #48)
