+++ appveyor.yml
@@ -73,6 +73,7 @@ build_script:
     - travis-tool.sh install_r tibble # workaround for MSnbase
     - travis-tool.sh install_bioc mzR xcms CAMERA Rdisop ropls BiocStyle Rgraphviz
     - travis-tool.sh install_github rickhelmus/patRoonData
+    - travis-tool.sh install_github blosloos/nontargetData
     - travis-tool.sh install_github blosloos/nontarget
     - travis-tool.sh install_github cbroeckl/RAMClustR
     - travis-tool.sh install_github rickhelmus/KPIC2
