### 2022-03-04 2a8fe4b
update machine image, try large resource_class [skip appveyor]
