### 2022-03-04 cfbec7d
group R dep installation together, possibly improving parallelization [skip ci]
