### 2022-03-04 #4029 @wesleyburr
@KristinaRiemer - I'm all done my review. This is a highly complex package, with many moving parts, so I have not verified all of the interaction and functionality, but the package installs, has some excellent thoughtful elements (e.g., the Docker setup), and has a tremendous breadth of applications. The help system is thorough and well written. I only came across one issue, which Rick immediately fixed up.
