+++ NEWS.md
@@ -19,6 +19,8 @@
 - Fixed: annotations slot for `featureGroups` was not updated when removing groups with `delete()` (except sets workflows)
 - Fixed: better handle missing spectrum data with spectrumSimilarity()
 - Made `nontarget` an optional dependency and install it from GitHub with CI and in the installation docs (see issue #48)
+- Fixed: MS files were not always correctly found
+
 
 # patRoon 2.0
 
