### 2022-03-07 943f5c7
properly handle pos/neg analyses changes
