### 2022-03-07 42102c8
small fixes for assertSuspectList()
