### 2022-03-07 bd8ea1c
handle non-numeric mz/rt input
