### 2022-03-07 59c1c4f
removed suspect screening columns from primary table to avoid repeated fGroup rows
