+++ NEWS.md
@@ -23,6 +23,7 @@
 - Fixed: `newProject()` ignored group/blank input for sets workflows
 - Fixed: better error handling for suspect lists with only one valid column
 - Improve suspect list handling when input `mz`/`rt` columns are not numeric
+- `checkFeatures()`: don't show multiple rows if a suspect was matched with multiple feature groups. This change removed the option to show specific suspect columns.
 
 
 # patRoon 2.0
