### 2022-03-07 #4029 @rickhelmus
Many thanks @KristinaRiemer!! I will have a look tomorrow morning. I will then also probably release a new minor version of patRoon with the fixes coming from the review issues and update the Zenodo links in the manuscript.
