### 2022-03-07 #4029 @KristinaRiemer
@rickhelmus I was going to ask you to do those tasks anyways, so sounds great! If you can provide the version number for the newest release and the DOI for the updated Zenodo archive, that would be great.
