### 2022-03-07 #4029 @KristinaRiemer
Thanks @wesleyburr & @hechth!

@rickhelmus I only found a couple of small errors in the paper to be fixed, as listed below:

- instances of "e.g." should be followed by a comma
- line 134: "is outlined **in** the project..."
- Depke et al. (2017) reference: possible to superscript the 2 in "MS2"?
- Krier et al. (2022) reference: capitalize L in Luxembourg in title

I'm not sure where the two DOIs listed by the bot above as missing or invalid are coming from? They don't appear to be in the paper, or anywhere in the code repo, that I can find?
