### 2022-03-08 8f71c63
small fix for fGroupPlotMode
