### 2022-03-08 41ce64e
Fixed: several issues with topMost plotting of EICs for sets data
