+++ NEWS.md
@@ -25,6 +25,7 @@
 - Improve suspect list handling when input `mz`/`rt` columns are not numeric
 - `checkFeatures()`: don't show multiple rows if a suspect was matched with multiple feature groups. This change removed the option to show specific suspect columns.
 - Fixed: `checkFeatures()` errored if Plot mode was 'Top most replicates' or 'All'
+- Fixed: several issues with `topMost` plotting of EICs for sets data
 
 
 # patRoon 2.0
