### 2022-03-08 dc2d1bb
ensure that peak filling works if plotting outside ylim
