### 2022-03-08 b625097
screenSuspects() with sets workflows: don't warn about set specific suspect data if all data is NA
