### 2022-03-08 6fea258
cleanup unavailable selections when saving the session
