### 2022-03-08 4d44e2b
Don't try to report TP components if no data is available
