+++ inst/report/TPs.Rmd
@@ -20,131 +20,134 @@ if (isScreening(rmdVars$fGroups))
 
 plotPathFull <- getPlotPath(FALSE)
 
-message("Generating parent plots...")
-prog <- openProgBar(0, nrow(cInfo))
-
-makeStructPlot <- function(SMI, out)
+if (length(components) > 0)
 {
-    mol <- getMoleculesFromSMILES(SMI, emptyIfFails = TRUE)[[1]]
-    withr::with_png(out, width = 1.5, height = 1.5, units = "in", res = 72, bg = NA, code = {
-        withr::with_par(list(mar = rep(0, 4)), plot(getRCDKStructurePlot(mol, 150, 150)))
-    })
-}
-
-plotIntMainArgs <- list(average = TRUE, normFunc = max)
-if (isFGSet(rmdVars$fGroups))
-{
-    plotIntMainArgs <- c(plotIntMainArgs, list(sets = TRUE))
-} else
-{
-    plotIntMainArgs <- c(plotIntMainArgs, list(col = "black"))
-}
-
-parentPlots <- setNames(Map(split(cInfo, seq_len(nrow(cInfo))), seq_len(nrow(cInfo)), f = function(parentRow, i)
-{
-    grpi <- match(parentRow$parent_group, names(rmdVars$fGroups))
-    plots <- list()
-    fg <- rmdVars$fGroups[, parentRow$parent_group]
+    message("Generating parent plots...")
+    prog <- openProgBar(0, nrow(cInfo))
     
-    plots[["int"]] <- file.path(plotPathFull, sprintf("int-parent_%d.png", grpi))
-    makeCachedPlot(plots[["int"]], "plotInt", c(list(fg), plotIntMainArgs), 3.3, 3.3,
-                   bg = NA, cacheDB = rmdVars$cacheDB)
-
-    if (!is.null(parentRow[["parent_SMILES"]]))
+    makeStructPlot <- function(SMI, out)
     {
-        plots[["struct"]] <- file.path(plotPathFull, sprintf("struct-parent_%d.png", grpi))
-        makeStructPlot(parentRow$parent_SMILES, plots[["struct"]])
-    }
-
-    setTxtProgressBar(prog, i)
-    
-    return(plots)
-}), cInfo$name)
-
-message("Generating TP plots...")
-prog <- openProgBar(0, length(components))
-cmpTab <- as.data.table(components)
-# isFGScrAnnotated <- isScreening(rmdVars$fGroups) && screenInfo(rmdVars$fGroups)[[""]]
-
-TPPlotName <- function(cmp, grp) paste0(cmp, "_", grp)
-
-TPPlots <- setNames(Map(split(cmpTab, seq_len(nrow(cmpTab))), seq_len(nrow(cmpTab)), f = function(ctRow, i)
-{
-    grpi <- match(ctRow$group, names(rmdVars$fGroups))
-    plots <- list()
-    fg <- rmdVars$fGroups[, ctRow$group]
-    
-    plots[["int"]] <- file.path(plotPathFull, sprintf("int-TP_%d.png", grpi))
-    makeCachedPlot(plots[["int"]], "plotInt", c(list(fg), plotIntMainArgs), 3.3, 3.3, bg = NA,
-                   cacheDB = rmdVars$cacheDB)
-
-    SMI <- ctRow[["SMILES"]]
-    if (!is.null(SMI))
-    {
-        plots[["struct"]] <- file.path(plotPathFull, sprintf("struct-parent_%d.png", grpi))
-        makeStructPlot(SMI, plots[["struct"]])
+        mol <- getMoleculesFromSMILES(SMI, emptyIfFails = TRUE)[[1]]
[…]
