### 2022-03-08 0bc1788
don't try to make formula comparison MS plots if candidate lacks MS/MS data
