### 2022-03-08 c715a88
temporarily disable SIRIUS compounds tests
