+++ paper.bib
@@ -5,8 +5,8 @@
   year  = 2021,
   publisher = {Zenodo},
   version = {2.0.0},
-  doi = {10.5281/zenodo.5752071},
-  url = {https://doi.org/10.5281/zenodo.5752071}
+  doi = {10.5281/zenodo.6337883},
+  url = {https://doi.org/10.5281/zenodo.6337883}
 }
 
 @article{Farre2008,
@@ -117,7 +117,7 @@
 }
 
 @article{Krier2022,
-  title = {Discovering pesticides and their TPs in Luxembourg waters using open cheminformatics approaches},
+  title = {Discovering pesticides and their TPs in {Luxembourg} waters using open cheminformatics approaches},
   journal = {Environment International},
   volume = {158},
   pages = {106885},
@@ -299,7 +299,8 @@
   number={1},
   pages={31--36},
   year={1988},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi = {10.1021/ci00057a005}
 }
 
 @article{Scholle2015,
@@ -434,7 +435,7 @@
   volume = {1071},
   pages = {19--28},
   author = {Tobias Depke and Raimo Franke and Mark Br\"{o}nstrup},
-  title = {Clustering of {MS}2 spectra using unsupervised methods to aid the identification of secondary metabolites from Pseudomonas aeruginosa},
+  title = {Clustering of {MS$^{2}$} spectra using unsupervised methods to aid the identification of secondary metabolites from Pseudomonas aeruginosa},
   journal = {Journal of Chromatography B}
 }
 
