### 2022-03-08 #4029 @rickhelmus
Hi @KristinaRiemer,

Thanks! I just pushed the manuscript corrections and released a new version of patRoon (2.0.1). The Zenodo link in the manuscript is now updated. The missing DOI was indeed missing and I added it to the references. The invalid DOI is actually not invalid, my guess is that the [parentheses in the DOI](https://github.com/rickhelmus/patRoon/blob/4ca0dcec9afb0f4808aec610197f6e0bb0f61d09/paper.bib#L361) are adding confusion. The link works fine in the manuscript. If it is a problem then perhaps I could change it to a DOI from https://shortdoi.org/ ?

Thanks,
Rick
