+++ R/TP-CTS.R
@@ -1,21 +1,20 @@
 #' @include main.R
-#' @include TP.R
+#' @include TP-structure.R
 NULL
 
 #' @export
-transformationProductsCTS <- setClass("transformationProductsCTS", contains = "transformationProducts")
+transformationProductsCTS <- setClass("transformationProductsCTS", contains = "transformationProductsStructure")
 
 setMethod("initialize", "transformationProductsCTS",
           function(.Object, ...) callNextMethod(.Object, algorithm = "cts", ...))
 
 # NOTE: this function is called by a withProg() block, so handles progression updates
-runCTS <- function(SMILES, transLibrary, generationLimit)
+runCTS <- function(SMILES, transLibrary, generationLimit, errorRetries)
 {
     CTSURL <- "https://qed.epa.gov/cts/rest/metabolizer/run"
     # CTSURL <- "https://qed.epacdx.net/cts/rest/metabolizer/run" # older version??
     
-    # UNDONE: args for parameters and retries
-    resp <- httr::RETRY("POST", url = CTSURL, encode = "json",
+    resp <- httr::RETRY("POST", url = CTSURL, encode = "json", times = errorRetries,
                         body = list(structure = SMILES, generationLimit = generationLimit,
                                     transformationLibraries = list(transLibrary)))
     
@@ -26,11 +25,11 @@ runCTS <- function(SMILES, transLibrary, generationLimit)
         return(NULL)
     
     cont <- httr::content(resp, "parsed")
-    if (cont$data$total_products == 0)
+    if (is.null(cont) || is.null(cont[["data"]]) || cont$data$total_products == 0)
         return(NULL)
 
     curTPID <- 0
-    processChilds <- function(chi, parentID = 0)
+    processChilds <- function(chi, parentID = 1)
     {
         curTPID <<- curTPID + 1
         res <- c(list(ID = curTPID, parentID = parentID), chi$data)
@@ -49,8 +48,8 @@ runCTS <- function(SMILES, transLibrary, generationLimit)
 
 
 #' @export
-generateTPsCTS <- function(parents, transLibrary, generationLimit = 1, skipInvalid = TRUE, fpType = "extended",
-                           fpSimMethod = "tanimoto", parallel = TRUE)
+generateTPsCTS <- function(parents, transLibrary, generationLimit = 1, errorRetries = 3, skipInvalid = TRUE,
+                           calcSims = FALSE, fpType = "extended", fpSimMethod = "tanimoto", parallel = TRUE)
 {
     checkmate::assert(
         checkmate::checkClass(parents, "data.frame"),
@@ -67,25 +66,31 @@ generateTPsCTS <- function(parents, transLibrary, generationLimit = 1, skipInval
                                             "photolysis_ranked", "mammalian_metabolism",
                                             "combined_abioticreduction_hydrolysis", 
                                             "combined_photolysis_abiotic_hydrolysis"), add = ac)
-    checkmate::assertCount(generationLimit, positive = TRUE, add = ac)
-    checkmate::assertFlag(skipInvalid, add = ac)
+    aapply(checkmate::assertCount, . ~ generationLimit + errorRetries, positive = TRUE, fixed = list(add = ac))
+    aapply(checkmate::assertFlag, . ~ skipInvalid + calcSims + parallel, fixed = list(add = ac))
     aapply(checkmate::assertString, . ~ fpType + fpSimMethod, min.chars = 1, fixed = list(add = ac))
-    checkmate::assertFlag(parallel, add = ac)
     checkmate::reportAssertions(ac)
     
     parents <- getTPParents(parents, skipInvalid)
     
-    baseHash <- makeHash(skipInvalid, fpType, fpSimMethod)
-    setHash <- makeHash(parents, baseHash)
-
-    # UNDONE: patRoon:::
[…]
