+++ DESCRIPTION
@@ -137,8 +137,8 @@ Collate:
     'main.R'
     'workflow-step.R'
     'TP.R'
-    'TP-CTS.R'
     'TP-structure.R'
+    'TP-CTS.R'
     'TP-biotransformer.R'
     'TP-library.R'
     'features.R'
