### 2022-03-09 #4029 @KristinaRiemer
The JOSS bot has been fixed to accept DOIs with parentheses now, thanks to @xuanxu! All of the DOIs in the paper are good now, thanks @rickhelmus.
