### 2022-03-09 caaa55d
add CTS properties
