+++ R/components-tps.R
@@ -212,7 +212,8 @@ doGenComponentsTPs <- function(fGroups, fGroupsTPs, ignoreParents, TPs, MSPeakLi
                 # limit columns a bit to not bloat components too much
                 # UNDONE: column selection OK?
                 prodCols <- c("name", "SMILES", "InChI", "InChIKey", "formula", "CID", "mass", "retDir", "trans_add",
-                              "trans_sub", "deltaMZ")
+                              "trans_sub", "deltaMZ", "generation", "likelihood", "accumulation", "production",
+                              "globalAccumulation")
                 prods <- prods[, intersect(names(prods), prodCols), with = FALSE]
                 
                 comps <- rbindlist(sapply(parentFGs, function(parentFG)
