+++ inst/report/TPs.Rmd
@@ -282,8 +282,9 @@ TPsDT <- rbindlist(Map(cTable, seq_along(cTable), names(cTable), f = function(cm
     
     splitCmp <- split(cmp, seq_len(nrow(cmp)))
     
-    trCols <- intersect(c("formula", "retDir", "TP_retDir", "retDiff", "mzDiff", "formulaDiff", "set", "CID",
-                          "SMILES"),
+    trCols <- intersect(c("formula", "retDir", "TP_retDir", "retDiff", "mzDiff", "formulaDiff",
+                          "generation", "likelihood", "accumulation", "production", "globalAccumulation",
+                          "set", "CID", "SMILES"),
                         names(cmp))
     ret[, TP := sapply(splitCmp, function(cr)
     {
