### 2022-03-14 #4029 @rickhelmus
Super happy to see it online now!! Thanks again @hechth, @wesleyburr, @KristinaRiemer! The submission led to a very nice and useful experience with JOSS.
